==> scripts/preliminary scripts/2__Rscripts_models_workflow.R <==
#######################################################################################
### Chapter 3-parasites and flocking species                                     ###
### Part 1 Models workflow                                                                       ###
### R-code                                                                          ###
### Jenny Munoz      
###
### Last update: March 2023                                                ###
################################################################################


# 0.Libraries  --------------------------------------------------------------

# libraries for easier manipulation of data
#install.packages("tidyr") 
install.packages("tidyverse") 
install.packages("dplyr")
install.packages ("data.table")
install.packages ("extrafont")
installed.packages("lubridate")  #for dates
#data cleaning
install.packages ("janitor")
install.packages("assertr")

#Other libraries for data analyses and visualizations
install.packages("vegan")
install.packages("ggplot2")
install.packages("devtools")
install.packages("knitr")
install.packages("ts")
install.packages("RColorBrewer")
install.packages("ggridges")
install.packages("ggtree")
install.packages("aplot")

#for models
install.packages("car") #Anova command
install.packages("lattice") #preliminary plots
install.packages("lme4") #for glmer (generalized linear mixed models) 
install.packages("visreg")  #extract confidence intervals and trend lines from GLMMs
install.packages("lsmeans") #least squared means
install.packages("MuMIn") #pseudo R squared for GLMMs
install.packages("emmeans")
#model assumptions
install.packages("DHARMa")

# bayes models 
install.packages('tidybayes')
install.packages ('bayesplot')
install.packages('brms') # bayesian approach to model phylogenetic data with repides observations
#install.packages('brmstools')
remotes::install_github("Pakillo/DHARMa.helpers")
devtools::install_github("mvuorre/brmstools")


# install.packages("remotes") # DHARMA FOR BRMS 

# Phylogenetic component
install.packages("ape")
#install.packages("here")
install.packages("phytools")
install.packages("tidyverse")
install.packages("metafor")
install.packages("phangorn") # to reconstruct a maximum clade credibility tree
install.packages("rr2")
install.packages ( "MCMCglmm")
#install.packages("phyr")
remotes::install_github("daijiang/phyr")
#options(repos = c(
 # phyr = 'https://daijiang.r-universe.dev',
 # CRAN = 'https://cloud.r-project.org'))
#install.packages('phyr')
install.packages("TreeTools")

# Libraries 
library(TreeTools)
#install.packages("INLA",repos=c(getOption("repos"),INLA="https://inla.r-inla-download.org/R/stable"), dep=TRUE) # to be able to run bayesian inference on the PGLMM models more info here https://www.r-inla.org
library("devtools")
#devtools::install_github(repo = "https://github.com/hrue/r-inla", ref = "stable", subdir = "rinla", build = FALSE)
#Libraries for data
library(tidyverse)
#library(tidyr)
#library(plyr)
library(ggplot2)
library(dplyr)
library(data.table)
library(extrafont)
library(lubridate)
# data cleaning 
library(janitor)
library(assertr)
# for data visualization
library(vegan)
library(ggplot2)
library(devtools)
library(knitr)
library(ts)
library(RColorBrewer)
library(ggridges)
library(ggtree)
library(aplot)

#libraries for models and visualizations
library(lattice) #preliminary plots
library(car) #Anova command
library(lsmeans) #least squared means
library(lme4) #for glmer (generalized linear mixed models) 
library(visreg) #extract confidence intervals and trend lines from GLMMs
library(MuMIn) #pseudo R squared for GLMMs
library(emmeans)
#mode assumption check
library (DHARMa)
#library(dplyr) 
#Phylogenetic component
library(ape)
#library(here)
library(phyr)
library(phytools)
library(metafor)
library (phangorn) # to reconstruct a maximum clade credibility tree
library (rr2)
library (MCMCglmm)
library(tidyverse)
library(skimr)
library(TreeTools)
#Libraries for plots
library(gridExtra)
library(ggpubr)
library(grid)
#libaries for bayes 
library(bayesplot)
library(tidybayes)
library(brms) # bayesian approach to model phylogenetic data with repides observations
library(DHARMa.helpers)
library(brmstools)



# # 1.Data import-----------------------------------------------------------------
#DATA
# This dataset contains all parasites samples (887) after removing exact duplicated rows, for which we have an assigned elevation, for 783 in total out of 998 that we had originally  (this included some duplicates)
# date preparation including date
#ectos_df<-read.csv("data/data_manuscript/7.dff_all_ectos_prevalence_abundance_individual_elevation_FILE.csv", na.strings =c("","NA")) %>% 
# rename(elevation=elevation_extrapolated_date) %>%
#  mutate(date=as.Date(full_date, format = "%Y-%m-%d")) %>% 
#  mutate(year_seasonality = format(date, "%j")) # # create new column for julian day yearly

#write.csv(ectos_df,"data/data_manuscript/7.dff_all_ectos_prevalence_abundance_individual_elevation_FILE.csv" )

ectos_birds_dff<-read.csv("data/data_manuscript/7.dff_all_ectos_prevalence_abundance_individual_elevation_FILE.csv", na.strings =c("","NA")) %>% 
  select(elevation, species_jetz, Powder.lvl,ectoparasites_PA, foraging_cat,sociality, total_lice,total_no_feathers_mites,total_mesostigmatidae,year_seasonality ) %>% 
  filter(species_jetz!="Premnoplex_brunnescens")


unique(ectos_df$species_jetz) # this is the total species that we have samples for 

phylo<-read.nexus("data/phylo_data/consensus/1_consensus_birdtreeManu_ectos_prevalence.nex")  # Need to prunne the tree 
unique(ectos_df$species_jetz )

# phylogenetic correlation structure, create a covariance matrix of species
phy_cov<-ape::vcv(phylo, corr=TRUE)


# ## 2.Data overview -------------------------------------------------------
str( ectos_df)
str(phylo)


# #### 3.Descriptive statistics --------------------------------------------------

# #### 4.Descriptive statistiscs plots -------------------------------------

phylo<-read.nexus("data/phylo_data/consensus/1_consensus_birdtreeManu_ectos_prevalence.nex")  # This include speceis form manu and iquitos  so need to rpun the tree in the data processin section

ectos_birds_dff<-read.csv("data/data_manuscript/7.dff_all_ectos_prevalence_abundance_individual_elevation_FILE.csv", na.strings =c("","NA")) %>% 
  select(elevation, species_jetz, Powder.lvl,ectoparasites_PA, foraging_cat,sociality, total_lice,total_no_feathers_mites,total_mesostigmatidae,date ) %>% 
  filter(species_jetz!="Premnoplex_brunnescens")

# just keeping the prevalences 
ectos_birds_dff_PA<-read.csv("data/data_manuscript/7.dff_all_ectos_prevalence_abundance_individual_elevation_FILE.csv", na.strings =c("","NA")) %>% 
  rename(elevation=elevation_extrapolated_date) %>%
  select( species_jetz, ectoparasites_PA, sociality ) %>% 
  filter(species_jetz!="Premnoplex_brunnescens")

unique (ectos_birds_dff_PA$species_jetz)

str(ectos_birds_dff_PA)
ectos_birds_dff_PA$ectoparasites_PA<-as.numeric(ectos_birds_dff_PA$ectoparasites_PA)
ectos_birds_dff_PA$species_jetz<-as.factor(ectos_birds_dff_PA$species_jetz)

# creating summaries per species
ectos_birds_dff_PA_species<-ectos_birds_dff_PA %>% group_by(species_jetz) %>% 
  summarise(total_presences=sum(ectoparasites_PA), sample_size=n(), sociality=max(sociality)) %>% 
  mutate(ectos_prevalence=total_presences/sample_size) %>% 
  filter(species_jetz!="Premnoplex_brunnescens")


ectos_birds_dff_mean <- ectos_birds_dff %>% select(species_jetz,total_lice,total_no_feathers_mites,total_mesostigmatidae) %>% filter(complete.cases(.)) %>% group_by(species_jetz) %>% mutate(mean_lice = mean(total_lice), mean_nf_mites=mean(total_no_feathers_mites)) 


### double check that the phylo and data match 
a<-(as.data.frame(phylo$tip.label))%>% mutate(name=phylo$tip.label) %>% select(name) %>% arrange(desc(name))
b<-(as.data.frame(ectos_birds_dff$species_jetz)) %>% mutate(name=ectos_birds_dff$species_jetz) %>% select(name) %>% arrange(desc(name)) %>% distinct(name)
tip<-as.list(setdiff(a,b))
print(tip)
# Drop some tips USE IF NEED TO DROP SOME TIPS when using the full phylogeny
phylo<-drop.tip (phylo, tip$name)

order<-(as.data.frame(phylo$tip.label)) # list of specie sin phylogenetic order
# Important!!!!!Make sure the names  are in the same order in the phylogeny and in the traits 
#rownames(ectos_birds_dff) <- ectos_birds_dff$species_jetz # first make it the row names 
#ectos_birds_dff<- ectos_birds_dff[match(phylo$tip.label,rownames(ectos_birds_dff)),]


# releveling
#plot_data <- data %>% select(genus_species, mite_load, ode_mass_g) %>% filter(complete.cases(.)) %>% group_by(genus_species) %>% mutate(avg_mass = mean(ode_mass_g)) %>% droplevels()
#plot_dat$genus_species <- as.factor(plot_dat$genus_species)
#plot_dat <- as.data.frame(plot_dat) %>% mutate(genus_species = fct_relevel(genus_species, plot_tree$tip.label[ordered_tips]))

# it will be nice to include the mean in this figures 
lice_load_plot <- ggplot(ectos_birds_dff, aes(x = total_lice, y =species_jetz)) +
  coord_cartesian(clip = "off") +
  geom_jitter(alpha=0.4, col="darkcyan")+
  scale_x_continuous(breaks=c(0, 1, 5, 10,15, 20, 30,40, 50)) +
  geom_point(data=ectos_birds_dff_mean, shape=124, size=1.5,aes(y=species_jetz, x=mean_nf_mites))+
  #scale_x_continuous(trans="log1p",breaks=c(0, 1, 2,5, 10, 25, 50, 100, 200, 300)) +
  scale_y_discrete(limits=order$`phylo$tip.label`) +
  ylab("") + xlab("Lice per individual") +
  theme_ridges(center_axis_labels = TRUE) +
theme_classic(10)+
  theme(panel.grid.major.y = element_line( size=.05, color="grey"))

lice_load_plot[["data"]][["species_jetz"]]

mites_load_plot <- ggplot(data = ectos_birds_dff, aes(y=species_jetz, x=total_no_feathers_mites)) + 
  geom_jitter(alpha=0.4, col="olivedrab") + #geom_boxplot(outlier.alpha=0) +
  geom_point(data=ectos_birds_dff_mean, shape=124, size=1.5,aes(y=species_jetz, x=mean_nf_mites))+
  scale_x_continuous(breaks=c(0, 1, 5, 10, 20, 30,40, 50, 100, 200, 300)) +
  scale_y_discrete(limits=order$`phylo$tip.label`) +
  theme_ridges(center_axis_labels = TRUE) + 
  ylab("") +
  xlab(" Mites per individual (Non_feather mites)") +
  theme_classic(10)+
  theme( axis.text.y=element_blank(),
                axis.title.y=element_blank(),
         panel.grid.major.y = element_line( size=.05, color="grey"))


#  remove this part to make sure teh species are in the same orderto eliminate the species names in the axes once we know the order is correct use
theme(
  axis.text.y=element_blank(),
  axis.title.y=element_blank())


# The phylogenetic plot with prevalence 

#tree_plot <- ggtree(phylo, ladderize=FALSE) + geom_tiplab() + ggplot2::xlim(0, 450)

ColorPalette <- brewer.pal(n = 9, name = "YlGnBu")
list.names=setNames(ectos_birds_dff_PA_species$ectos_prevalence, ectos_birds_dff_PA_species$species_jetz)
fmode<-as.factor(setNames(ectos_birds_dff_PA_species$sociality,ectos_birds_dff_PA_species$species_jetz))
tree_plot_ectos<- contMap(phylo, list.names, direction = "leftwards", plot=FALSE)
#object_color<-setMap(object, c("snow3","darkslategray3","dodgerblue","darkolivegreen3","goldenrod1"))
object_color<-setMap(tree_plot_ectos, ColorPalette)
tree_plot_sociality<-dotTree(phylo,fmode,colors=setNames(c("yellow","black"), c("1","0")),ftype="i",fsize=0.5, lwd=4) 
plot(tree_plot_sociality)
# 
png("figures/figures_manuscript/Fig1.PhyloTree_prevalence.png", width = 2500, height = 3100, res = 300, units = "px")
plot(dotTree(phylo,fmode,colors=setNames(c("coral3","black"),c("1","0")),ftype="i",fsize=0.5, lwd=4),text(x=10,y=-5,"Mixed-species flocks",pos=1))
plot(object_color$tree,colors=object_color$cols,add=TRUE,ftype="off",lwd=5,fsize=0.5,
     xlim=get("last_plot.phylo",envir=.PlotPhyloEnv)$x.lim,
     ylim=get("last_plot.phylo",envir=.PlotPhyloEnv)$y.lim)
add.color.bar(10, object_color$cols, title = "", lims = object$lims, digits = 3, prompt=FALSE,x=70,y=-5, lwd=4,fsize=1,subtitle="Ectoparasites Prevalence",pos=4)
dev.off()

unique( ectos_birds_dff_PA_species$species_jetz)

phylo_mite_lice_plot <- lice_load_plot %>% insert_right(mites_load_plot) 


#ggsave("figures/figures_manuscript/Fig1b__mite_lice_plot.png", plot=phylo_mite_lice_plot, height=10, width=12, units="in")



# Plotting the traits things in the  phylogeny 
#Some examples fr plotting http://www.phytools.org/Cordoba2017/ex/15/Plotting-methods.html
# Plot phylogenetic tree and  the trait continous trait ( prevalence)
#ContMap #Function plots a tree with a mapped continuous character. 
#The mapping is accomplished by estimating states at internal nodes using ML with fastAnc, and then interpolating the states along each edge using equation [2] of Felsenstein (1985).

#contMap(tree, x, res=100, fsize=NULL, ftype=NULL, lwd=4, legend=NULL,
#lims=NULL, outline=TRUE, sig=3, type="phylogram", direction="rightwards", 
#plot=TRUE, ...)


ectos_pres_abs<-ectos_birds_dff %>% group_by(species_jetz ) %>% 
  summarise(ectoparasites_presence=(sum(ectoparasites_PA)), sample_size=(n()), elevation_midpoint=max(elevation_midpoint)) %>% 
  mutate(proportion_ectoparasites=ectoparasites_presence/sample_size)

species_atributes<-ectoparasites_df %>% select( species_jetz, species_clean,species_binomial,BLFamilyLatin, sociality,mass_tidy,elevation_cat, foraging_cat )
species_attributes_distict<-distinct(species_atributes)

ectos_pres_abs_df<-right_join(species_attributes_distict, ectos_pres_abs, by="species_jetz")   %>% arrange(elevation_cat)

ectos_pres_abs_df<-ectos_pres_abs_df %>% distinct(species_jetz,proportion_ectoparasites,.keep_all = TRUE) # Eliminates species taht ocurr at two elevatiosn and keep only one ( in alphabetical order of the stations, e.g for galbula low_elevations is kept and montane is deleted, but the samples size is mantained this is jut to be able to do the phylogenetic analyses properly)

ectos_pres_abs_df<- ectos_pres_abs_df %>% rename(family=BLFamilyLatin)


ectos_birds_dff<-read.csv("data/data_manuscript/7.dff_all_ectos_prevalence_abundance_individual_elevation_FILE.csv", na.strings =c("","NA")) %>% 
  rename(elevation=elevation_extrapolated_date) %>%
  select(elevation, species_jetz, Powder.lvl,ectoparasites_PA, foraging_cat,sociality) 

ectos_birds_dff_PA<-ectos_birds_dff %>% group_by(species_jetz ) %>% 
  summarise(ectoparasites_presence=(sum(ectoparasites_PA)), sample_size=(n()), elevation_mean=mean(elevation), sociality=max(sociality))%>% 
  mutate(proportion_ectoparasites=ectoparasites_presence/sample_size)

unique (ectos_birds_dff_PA$species_jetz )
#ectos_birds_dff_mean <- ectos_birds_dff %>% select(species_jetz,total_lice,total_no_feathers_mites,total_mesostigmatidae) %>% filter(complete.cases(.)) %>% group_by(species_jetz) %>% mutate(mean_lice = mean(total_lice), mean_nf_mites=mean(total_no_feathers_mites)) %>% droplevels()
phylo<-read.nexus("data/phylo_data/consensus/1_consensus_birdtreeManu_ectos_prevalence.nex")  # This include specis form manu and iquitos  so need to rpun the tree in the data processin section

a<-(as.data.frame(phylo$tip.label))%>% mutate(name=phylo$tip.label) %>% select(name) %>% arrange(desc(name))
b<-(as.data.frame(ectos_birds_dff$species_jetz)) %>% mutate(name=ectos_birds_dff$species_jetz) %>% select(name) %>% arrange(desc(name)) %>% distinct(name)

tip<-as.list(setdiff(a,b))
print(tip)

# Drop some tips USE IF NEED TO DROP SOME TIPS when using the full phylogeny
phylo<-drop.tip (phylo, tip$name) 


# ##### 5.Data processing prevalence ectos ----------------------------------------------------
ectos_birds_dff<-read.csv("data/data_manuscript/7.dff_all_ectos_prevalence_abundance_individual_elevation_FILE.csv", na.strings =c("","NA")) %>% 
  select(elevation, species_jetz, Powder.lvl,ectoparasites_PA, foraging_cat,sociality, year_seasonality ) %>% 
  na.omit() 
phylo<-read.nexus("data/phylo_data/consensus/1_consensus_birdtreeManu_ectos_prevalence.nex")  # Need to prunne the tree 

#ectos_birds_dff <-ectos_birds_dff  %>% mutate_at("species_jetz", str_replace, " ", "_")
#ectos_birds_dff$elevation_cat<-as.factor(ectos_birds_dff$elevation_cat)
ectos_birds_dff$foraging_cat<-as.factor(ectos_birds_dff$foraging_cat)
ectos_birds_dff$species_jetz<-as.factor(ectos_birds_dff$species_jetz)
ectos_birds_dff$elevation<-as.numeric(ectos_birds_dff$elevation)
ectos_birds_dff$year_seasonality<-as.numeric(ectos_birds_dff$year_seasonality)
#ectos_birds_dff$elevation_midpoint<-as.numeric(ectos_birds_dff$elevation_midpoint)
ectos_birds_dff$sociality<-as.factor(ectos_birds_dff$sociality)
ectos_birds_dff$Powder.lvl<-as.factor(ectos_birds_dff$Powder.lvl)
ectos_birds_dff$ectoparasites_PA<-as.numeric(ectos_birds_dff$ectoparasites_PA)
ectos_birds_dff$species<- ectos_birds_dff$species_jetz # create a column for the species effect different to the phylogenetic one

names(ectos_birds_dff)
is.ultrametric(phylo)

# Make sure the tips and the names on the file coincide and formatting of name is consistent
phylo$edge.length  
phylo$tip.label
is.binary(phylo)

# Make sure this two are the same numbers 
a<-(as.data.frame(phylo$tip.label))%>% mutate(name=phylo$tip.label) %>% select(name) %>% arrange(desc(name))
b<-(as.data.frame(ectos_birds_dff$species_jetz)) %>% mutate(name=ectos_birds_dff$species_jetz) %>% select(name) %>% arrange(desc(name)) %>% distinct(name)

tip<-as.list(setdiff(a,b))
print(tip)

# Drop some tips USE IF NEED TO DROP SOME TIPS when using the full phylogeny
phylo<-drop.tip (phylo, tip$name) 

phy_cov<-ape::vcv(phylo, corr=TRUE)

# ##### 5.1.Analyses models prevalence ectos --------------------------------------------------------
###_###_###
  #a) model glmm
ectos_p_glmm <-glmer(ectoparasites_PA~sociality+scale(elevation)+(1|Powder.lvl)+(1|species_jetz), #+elevation_midpoint+Powder.lvl
                             data = ectos_birds_dff, 
                             family = "binomial")
###_###_###
  #b) model pglmm
ectos_p_pglmm <-phyr::pglmm(ectoparasites_PA~sociality+scale(elevation)+(1|species_jetz__)+(1|Powder.lvl), #+elevation_midpoint+Powder.lvl +(1|Powder.lvl)
                                    data = ectos_birds_dff, # exlude raws with nas this is usefull to check the model assumptions 
                                    family = "binomial",
                                    cov_ranef = list(species_jetz= phylo), #class phylo
                                    #bayes = TRUE,
                                    REML = TRUE, 
                                    verbose = TRUE,
                                    s2.init = .25) # what is this last parameter for
###_###_###
    #c) model pglmm summary
names(ectos_birds_dff)
class(ectos_p_pglmm ) 
summary(ectos_p_pglmm )
print(ectos_p_pglmm )
fixef(ectos_p_pglmm )
predict(ecto_p_pglmm )
rr2::R2(ectos_p_pglmm ) # goodness of fit

    #model pglmm assumptions check DHARMa
    #Overdisperssion
simulationOutput <-DHARMa::simulateResiduals(fittedModel =ectos_p_pglmm, plot = F)
plot(simulationOutput)
plotResiduals(simulationOutput , ectos_p_pglmm$data$sociality)
plotResiduals(simulationOutput , form =ectos_p_pglmm$sociality)

    #Homogenity of variance within groups (Heteroscedasticity) 
testCategorical(simulationOutput, catPred = ectos_birds_dff$sociality)

    # Other test
testUniformity(simulationOutput) #tests if the overall distribution conforms to expectations # 
testOutliers(simulationOutput)#  tests if there are more simulation outliers than expected
testDispersion(simulationOutput) # tests if the simulated dispersion is equal to the observed dispersion
testQuantiles(simulationOutput) #fits a quantile regression or residuals against a predictor (default predicted value), and tests of this conforms to the expected quantile
testCategorical(simulationOutput, catPred = ectos_birds_dff$sociality)# tests residuals against a categorical predictor
testZeroInflation(simulationOutput) ## tests if there are more zeros in the data than expected from the simulations

###_###_###
    #d) model pglmm bayes : hierarchical Bayesian models fitted using integrated nested laplace approximation (INLA)
###_###_###
ecto_p_pglmm_bayes <- phyr::pglmm(ectoparasites_PA~sociality+scale(elevation)+(1|species_jetz__)+(1|Powder.lvl),
                                     data =ectos_birds_dff, 
                                     cov_ranef = list(species_jetz= phylo), #class phylo can take a phylohgeny or a covariance matrix
                                     family = "binomial", # zeroinflated.binomial in reality I want to use a zeroinflated.binomial but for some reason i can not plot those
                                     bayes = TRUE,
                                     prior = "pc.prior.auto")

summary(ecto_p_pglmm_bayes)
print(ecto_p_pglmm_bayes)
rr2::R2(ecto_p_pglmm_bayes) #0.1940
class(ecto_p_pglmm_bayes) 
# plot 
estimates_plot<-plot_bayes(ecto_p_pglmm_bayes)
png("data/data_analyses/models/model_plots/1.estimates_plot_model_prevalence_pglmm_phylo_multiple_obs_031623.png",width = 3000, height = 3000, res = 300, units = "px")
estimates_plot
dev.off()

# assumptions check model pglmm_ bayes DHARMa ( DHARMa is not working with pgllm Bayes = True)
res_bayes<-DHARMa::simulateResiduals(ecto_p_pglmm_bayes)
class(ecto_p_pglmm_bayes)

## Model Individual  sample with phylognetic + non-phylo effects 
#d) model BRMS bayes


ecto_p_brms_bayes<-brms::brm(ectoparasites_PA~
                               sociality+
                               scale(elevation)+
                               scale(year_seasonality)+
                               socality:scale(elevation)+
                               sociality:scale(year_seasonality)+
                               scale(elevation):scale(year_seasonality)+
                               sociality:scale(year_seasonality):scale(elevation)+
          (1|gr(species_jetz, cov = phy_cov))+  #(1|Powder.lvl)
            (1|Powder.lvl)+
            (1|species),
        data=ectos_birds_dff,
        family= bernoulli(), # bernoulli() uses the (link = "logit").#zero_inflated_negbinomial() 
        data2 = list(phy_cov=phy_cov),
        iter=6000, warmup=3000, #First we need the specify how many iteration we want the MCMC to run, We need to specify how many chains we want to run.
        thin=2,
        control=list(adapt_delta=0.99, max_treedepth=12)) 

loo_compare(b_ecto_p_brms_bayes,ecto_p_brms_bayes)

m_no_interactions<-loo(b_ecto_p_brms_bayes)
m_interactions<-loo(m_no_interactions,m_interactions)
loo_compare(m_no_interactions,m_interactions)
saveRDS(ecto_p_brms_bayes, "data/data_analyses/models/seasonality/1.model_prevalence_brms_phylo_multiple_obs_031623_seasonality_inter_elev_season.RDS")
saveRDS(b_ecto_p_brms_bayes, "data/data_analyses/models/seasonality/1.model_prevalence_brms_phylo_multiple_obs_031623_seasonality.RDS")
b_ecto_p_brms_bayes<-readRDS("data/data_analyses/models/seasonality/1.model_prevalence_brms_phylo_multiple_obs_031623_seasonality.RDS")

#saveRDS(ecto_p_brms_bayes, "data/data_analyses/models/1.model_prevalence_brms_phylo_multiple_obs_031623.RDS")
#ecto_p_brms_bayes<-readRDS("data/data_analyses/models/1.model_prevalence_brms_phylo_multiple_obs_031623.RDS")
# Summarize the model
summary (ecto_p_brms_bayes)
levels(ectos_birds_dff$sociality)

fixef() # to get more detailed values for estimates
coef(ecto_p_brms_bayes) # if you have group-level effects (hierarchical data)

# interpret the model 
#To test whether all regression coefficients are different from zero, we can look at the Credible Intervals that are listed in the summary output or we can visually represent them in density plots.
#If we do so, we clearly see that zero is not included in any of the density plots, meaning that we can be reasonably certain the regression coefficients are different from zero.
#INTERPRETATION:In the model, the parameter for Sociality means the expected difference between non_social(0) and social (1) with all other covariates held constant. we clearly see that zero is included in the density plot for sociality so there is not effect of sociality??
bayes_R2(ecto_p_brms_bayes) # R2 0.1529
plot(ecto_p_brms_bayes) # paarameter distributio and convergence

mcmc_plot(ecto_p_brms_bayes) # Dots represent means of posterior distribution along with 95% CrIs, as estimated by the bmod5 model
launch_shinystan()
pp_check(ecto_p_brms_bayes, ndraws = 100)+ xlim(0, 5)  #  test for the model fit to the data .need to modify the scale of this plot posterior predictive checks, 100 random draws or distributions created by the model 
pp_check(ecto_p_brms_bayes, type="bars", ndraws = 100)+ xlim(0, 20) 

pp_m<- brms::posterior_predict(ecto_p_brms_bayes)
log1 <- scale_x_continuous(trans="log1p")
ppc_dens_overlay(y=ecto_p_brms_bayes$data$ectoparasites_PA, pp_m[1:200, ])  + 
  coord_cartesian(xlim = c(0, 10))

ppc_rootogram(y=ecto_p_brms_bayes$data$ectoparasites_PA, pp_m[1:200, ])  +   
  coord_cartesian(xlim = c(0, 5), ylim = c(0,30))

ppc_stat(y=ecto_p_brms_bayes$data$ectoparasites_PA, pp_m, stat = "prop_zero")

#Assumptions check model pglmm_ bayes
#remotes::install_github("Pakillo/DHARMa.helpers")

simulate_residuals <- dh_check_brms(ecto_p_brms_bayes, integer = TRUE)
  plot( simulate_residuals, form = ectos_birds_dff$sociality)
  DHARMa::testDispersion(simulate_residuals)
  DHARMa::testZeroInflation(simulate_residuals ) ## tests if there are more zeros in the data than expected from the simulations
  testUniformity(simulate_residuals) #tests if the overall distribution conforms to expectations
  
# assumptiosn look good!   
  

  
# Plots BRMS 
   color_scheme_set("blue")

  # convergence 
  png("data/data_analyses/models/model_plots/1.parameters_distribution_convergence_plot_model_PREVALENCE_brms_phylo_multiple_obs_032123.png",width = 3000, height = 3000, res = 300, units = "px")
  plot(ecto_p_brms_bayes)
  dev.off()
  
  # model fit
  png("data/data_analyses/models/model_plots/1.model_fit_PREVALENCE_LICE_brms_phylo_multiple_obs_032123.png.png",width = 3000, height = 3000, res = 300, units = "px")
  pp_check(ecto_p_brms_bayes, ndraws = 100)+ xlim(0, 5)  #  test for the model fit to the data .need to modify the scale of this plot posterior predictive checks, 100 random draws or distributions created by the model 
  dev.off()
  

  # Posterior distributions
  
color_scheme_set("green")
#brmstools::forest(ecto_p_brms_bayes, level = 0.95, show_data = TRUE)

estimates_plot<-mcmc_plot(ecto_p_brms_bayes ,prob=0.90, prob_outer=0.95,
                          variable = c("b_Intercept", "b_sociality1", "b_scaleelevation","sd_Powder.lvl__Intercept","sd_species__Intercept","sd_species_jetz__Intercept"),
                          type="areas") +
  labs(title="Posterior distributions ectoparasite prevalence", subtitle ="with medians and 95% intervals")+
  theme_minimal(20)+
  geom_vline(xintercept = 0, linetype = 2, colour = "grey40")+
  xlab("Estimate")

estimates_plot_intervals<-mcmc_plot(ecto_p_brms_bayes ,prob=0.90, prob_outer=0.95,point_est = "mean",
                          variable = c("b_Intercept", "b_sociality1", "b_scaleelevation","sd_Powder.lvl__Intercept","sd_species__Intercept","sd_species_jetz__Intercept"),
                          type="intervals") +
  labs(title="Posterior distribution ectoparasite prevalence", subtitle ="with means and 95% intervals")+
  theme_minimal(20)+
  geom_vline(xintercept = 0, linetype = 2, colour = "grey40")+
  xlab("Estimate")

png("data/data_analyses/models/model_plots/1.parameters_plot_model_PREVALENCE_brms_phylo_multiple_obs_032123.png",width = 3000, height = 3000, res = 300, units = "px")
estimates_plot
dev.off()


# ##### 5.2 Model predictions plots prevalence ectos----------------------------------
# Descriptive plot 

ectos_birds_dff<-read.csv("data/data_manuscript/7.dff_all_ectos_prevalence_abundance_individual_elevation_FILE.csv", na.strings =c("","NA")) %>% 
  rename(elevation=elevation_extrapolated_date) %>%
  select(elevation, species_jetz, Powder.lvl,foraging_cat, sociality, ectoparasites_PA) %>% 
  na.omit() %>% filter(species_jetz!="Premnoplex_brunnescens")  #Removing outliers for total mites

str(ectos_birds_dff)
ectos_birds_dff$ectoparasites_PA<-as.numeric(ectos_birds_dff$ectoparasites_PA)


# creating summaries per species
ectos_birds_dff_PA_species<-ectos_birds_dff_PA %>% group_by(species_jetz) %>% 
  summarise(total_presences=sum(ectoparasites_PA), sample_size=n(), sociality=max(sociality)) %>% 
  mutate(ectos_prevalence=total_presences/sample_size) %>% 
  filter(species_jetz!="Premnoplex_brunnescens")

ectos_birds_dff_PA_summary<-ectos_birds_dff_PA_species %>% group_by(sociality) %>% 
  summarize( mean=mean(ectos_prevalence))

ectos_birds_dff_mean <- ectos_birds_dff %>% select(species_jetz,total_lice,total_no_feathers_mites,total_mesostigmatidae) %>% filter(complete.cases(.)) %>% group_by(species_jetz) %>% mutate(mean_lice = mean(total_lice), mean_nf_mites=mean(total_no_feathers_mites)) 


#filter(total_lice<60)  # removing outliers 


png("figures/figures_manuscript/Fig1_Ectoparasite_prevalence_data_plot.png", width = 3000, height = 3000, res = 300, units = "px")
ggplot(data = ectos_birds_dff_PA_species, aes(x = sociality, y=ectos_prevalence))+
  # geom_point(data = ectos_df, aes(x=sociality, y = proportion_ectoparasites),color="grey",size=2)+
  geom_jitter(data = ectos_birds_dff_PA_species, aes(x=sociality, y = ectos_prevalence),color="grey",size=3,width = 0.07)+
  geom_point(data =ectos_birds_dff_PA_summary, aes(x=sociality, y = mean),color="cyan4",size=5)+
  scale_y_continuous("Ectoparasite prevalence", limits = c(0,1)) +
  scale_x_discrete("Sociality", limits =c(0,1))+
  #geom_hline(yintercept = 0.775, linetype = "dashed")+
  theme_classic(40)
dev.off()




# ####### 6.Data processing abundance lice----------------------------------------------------
ectos_birds_dff<-read.csv("data/data_manuscript/7.dff_all_ectos_prevalence_abundance_individual_elevation_FILE.csv", na.strings =c("","NA")) %>% 
  rename(elevation=elevation_extrapolated_date) %>%
  select(elevation, species_jetz, Powder.lvl,total_lice,foraging_cat, sociality, total_lice) %>% 
  na.omit() %>% filter(species_jetz!="Premnoplex_brunnescens") 
  #filter(total_lice<60)  # removing outliers 

phylo<-read.nexus("data/phylo_data/consensus/1_consensus_birdtreeManu_ectos_prevalence.nex")  # This include speceis form manu and iquitos  so need to rpun the tree in the data processin section


str(ectos_birds_dff)
mean(as.numeric(ectos_birds_dff$total_lice), na.rm=T)
# Finding putliers
#ectos_birds_dff <-ectos_birds_dff  %>% mutate_at("species_jetz", str_replace, " ", "_")
#ectos_birds_dff$elevation_cat<-as.factor(ectos_birds_dff$elevation_cat)
ectos_birds_dff$foraging_cat<-as.factor(ectos_birds_dff$foraging_cat)
ectos_birds_dff$species_jetz<-as.factor(ectos_birds_dff$species_jetz)
ectos_birds_dff$elevation<-as.numeric(ectos_birds_dff$elevation)
#ectos_birds_dff$elevation_midpoint<-as.numeric(ectos_birds_dff$elevation_midpoint)
ectos_birds_dff$sociality<-as.factor(ectos_birds_dff$sociality)
ectos_birds_dff$Powder.lvl<-as.factor(ectos_birds_dff$Powder.lvl)
ectos_birds_dff$total_lice<-as.numeric(ectos_birds_dff$total_lice)
ectos_birds_dff$species<- ectos_birds_dff$species_jetz # create a column for the species effect different to the phylogenetic one
names(ectos_birds_dff)
is.ultrametric(phylo)

# Make sure the tips and the names on the file coincide and formating of name is consitent
phylo$edge.length  
phylo$tip.label
is.binary(phylo)

# Make sure this two are the same numbers 
a<-(as.data.frame(phylo$tip.label))%>% mutate(name=phylo$tip.label) %>% select(name) %>% arrange(desc(name))
b<-(as.data.frame(ectos_birds_dff$species_jetz)) %>% mutate(name=ectos_birds_dff$species_jetz) %>% select(name) %>% arrange(desc(name)) %>% distinct(name)

tip<-as.list(setdiff(a,b))
print(tip)

# Drop some tips USE IF NEED TO DROP SOME TIPS when using the full phylogeny
phylo<-drop.tip (phylo, tip$name) 

#Phylogenetic covariance matrix
phy_cov<-ape::vcv(phylo, corr=TRUE)

# ####### 6.1.Analyses models abundance lice--------------------------------------------------------
# Lice

###_###_###
#a) model glmm
###_###_###
 str(ectos_birds_dff)
lice_a_glmm <-glmer(total_lice~sociality+scale(elevation)+(1|Powder.lvl)+(1|species_jetz), #+elevation_midpoint+Powder.lvl
                    data = ectos_birds_dff, 
                    family = "poisson")

# Summary
summary(lice_a_glmm )
rr2::R2(lice_a_glmm)# R2 Predicted 0.3951
fixef(lice_a_glmm)
predict(lice_a_glmm)
class(lice_a_glmm ) 

# Assumptions check
simulationOutput_a_lice<- DHARMa::simulateResiduals(fittedModel =lice_a_glmm, plot = F,integerResponse = T, re.form = NULL ) #quantreg=T
plot(simulationOutput_a_lice)
testUniformity(simulationOutput_a_lice) #tests if the overall distribution conforms to expectations
testOutliers(simulationOutput_a_lice)#  tests if there are more simulation outliers than expected
testDispersion(simulationOutput_a_lice) # tests if the simulated dispersion is equal to the observed dispersion
testQuantiles(simulationOutput_a_lice) #fits a quantile regression or residuals against a predictor (default predicted value), and tests of this conforms to the expected quantile
testCategorical(simulationOutput_a_lice, catPred = ectos_birds_dff$sociality)# tests residuals against a categorical predictor
testZeroInflation(simulationOutput_a_lice) ## tests if there are more zeros in the data than expected from the simulations

# zero inflated, overdispersed, with outliers
#b) model pglmm
###_###_###

lice_a_pglmm <-phyr::pglmm(total_lice~sociality+scale(elevation)+(1|species_jetz__)+(1|Powder.lvl),
                                   data = ectos_birds_dff, 
                                   family = "poisson",
                                   cov_ranef = list(species_jetz= phylo), #class phylo
                                   #bayes = TRUE,
                                   add.obs.re = TRUE,
                                   REML = TRUE, 
                                   verbose = TRUE,
                                   s2.init = .25) # what is this last parameter for

histogram(ectos_birds_dff$total_lice) # id some outliers 
plot_data(lice_a_pglmm)
summary(lice_a_pglmm)
rr2::R2(lice_a_pglmm)
fixef(lice_a_pglmm)
predict(ecto_a_pglmm)

# Assumptions check
simulationOutput_a_lice_2<- DHARMa::simulateResiduals(fittedModel=lice_a_pglmm, plot = F,integerResponse = T, re.form = NULL ) #quantreg=T
plot(simulationOutput_a_lice_2)
testUniformity(simulationOutput_a_lice_2) #tests if the overall distribution conforms to expectations
testOutliers(simulationOutput_a_lice_2)#  tests if there are more simulation outliers than expected
testDispersion(simulationOutput_a_lice_2) # tests if the simulated dispersion is equal to the observed dispersion
testQuantiles(simulationOutput_a_lice_2) #fits a quantile regression or residuals against a predictor (default predicted value), and tests of this conforms to the expected quantile
testCategorical(simulationOutput_a_lice_2, catPred = ectos_birds_dff$sociality)# tests residuals against a categorical predictor
testZeroInflation(simulationOutput_a_lice_2) ## tests if there are more zeros in the data than expected from the simulations

# no outliers, but no homogenity of varianse KS and withig groups significant, and zero inflated

#c) model pglmm bayes: hierarchical Bayesian models fitted using integrated nested laplace approximation (INLA)
###_###_###

zip_lice_a_pglmm_bayes <-phyr::pglmm(total_lice~sociality+scale(elevation)+(1|species_jetz__)+(1|Powder.lvl),
                                         data =ectos_birds_dff, 
                                         family ="zeroinflated.poisson", #POISSON  ="zeroinflated.poisson", #
                                         cov_ranef = list(species_jetz= phylo), #class phylo
                                         bayes = TRUE,
                                         verbose=FALSE,
                                         prior = "inla.default") # consider using    add.obs.re = T
#saveRDS(zip_lice_a_pglmm_bayes, "data/data_analyses/models/2.model_ABUNDANCE_LICE_pglmm_zip_phylo_multiple_obs_17032023.RDS")
zip_lice_a_pglmm_bayes<-readRDS ("data/data_analyses/models/2.model_ABUNDANCE_LICE_pglmm_zip_phylo_multiple_obs_17032023.RDS")
#1) Summary of the model 
communityPGLMM.plot.re(x=zip_lice_a_pglmm_bayes ) 
summary(zip_lice_a_pglmm_bayes)   
rr2::R2(zip_lice_a_pglmm_bayes)
class(zip_lice_a_pglmm_bayes)

launch_shinystan(zip_lice_a_pglmm_bayes)

# Plots

estimates_plot<-plot_bayes(zip_lice_a_pglmm_bayes ) # for some reason does not allow me to plot zero inflated poisson 

#png("data/data_analyses/models/model_plots/2.parameters_plot_model_LICE_ABUNDANCE_pglmm_zip_phylo_multiple_obs_032123.png",width = 3000, height = 3000, res = 300, units = "px")
#estimates_plot
#dev.off()

# Assumptions check DHARMa does not work with pglm bayes=TRUE so I can not evaluate model fit.
simulationOutput_a_lice_3<- DHARMa::simulateResiduals(fittedModel=zip_lice_a_pglmm_bayes, plot = F,integerResponse = T, re.form = NULL ) #quantreg=T
plot(simulationOutput_a_lice_3)
testUniformity(simulationOutput_a_lice_3) #tests if the overall distribution conforms to expectations
testOutliers(simulationOutput_a_lice_3)#  tests if there are more simulation outliers than expected
testDispersion(simulationOutput_a_lice_3) # tests if the simulated dispersion is equal to the observed dispersion
testQuantiles(simulationOutput_a_lice_3) #fits a quantile regression or residuals against a predictor (default predicted value), and tests of this conforms to the expected quantile
testCategorical(simulationOutput_a_lice_3, catPred = ectos_birds_dff$sociality)# tests residuals against a categorical predictor
testZeroInflation(simulationOutput_a_lice_3) ## tests if there are more zeros in the data than expected from the simulations


#d) model pglmm bayes : zero_inflated_negbinomial, ind scaled elevation
###_###_###

zinb_lice_a_brms_bayes<-brm(total_lice~sociality+scale(elevation)+
                     (1|gr(species_jetz, cov = phy_cov))+  
                    (1|Powder.lvl) + 
                    (1|species),
                    data=ectos_birds_dff,
                    family=zero_inflated_negbinomial(),  #zero_inflated_negbinomial()
                     data2 = list(phy_cov=phy_cov),
                    iter=6000, warmup=3000,
                    thin=2,
                    control=list(adapt_delta=0.99, max_treedepth=12)) 
zinb_lice_a_brms_bayes<-lice_a_brms_bayes
#saveRDS(lzinb_lice_a_brms_bayes, "data/data_analyses/models/2.model_ABUNDANCE_LICE_brms_zinb_phylo_multiple_obs_17032023.RDS")
zinb_lice_a_brms_bayes<-readRDS("data/data_analyses/models/2.model_ABUNDANCE_LICE_brms_zinb_phylo_multiple_obs_17032023.RDS")


#  exploring zero inflated poisson 
 
zip_lice_a_brms_bayes<-brm(total_lice~sociality+scale(elevation)+
                              (1|gr(species_jetz, cov = phy_cov))+  
                              (1|Powder.lvl) + 
                              (1|species),
                            data=ectos_birds_dff,
                            family=zero_inflated_poisson(),  #zero_inflated_negbinomial()
                            data2 = list(phy_cov=phy_cov),
                            iter=6000, warmup=3000,
                            thin=2,
                            control=list(adapt_delta=0.99, max_treedepth=12)) 

zip_lice_a_brms_bayes<-lice_a_brms_bayes
#saveRDS(zip_lice_a_brms_bayes, "data/data_analyses/models/2.model_ABUNDANCE_LICE_brms_zip_phylo_multiple_obs_21032023.RDS")

# some outliers and smaller R2, overall performed porrly compared to zeroinflated negative binomial)_


# Summarize the model
summary(zinb_lice_a_brms_bayes)


fixef(zinb_lice_a_brms_bayes) # to get more detailed values for estimates
coef(zinb_lice_a_brms_bayes) # if you have group-level effects (hierarchical data)

# interpret the model 
#To test whether all regression coefficients are different from zero, we can look at the Credible Intervals that are listed in the summary output or we can visually represent them in density plots.
#If we do so, we clearly see that zero is not included in any of the density plots, meaning that we can be reasonably certain the regression coefficients are different from zero.
#INTERPRETATION:In the model, the parameter for Sociality means the expected difference between non_social(0) and social (1) with all other covariates held constant. we clearly see that zero is included in the density plot for sociality so there is not effect of sociality??
bayes_R2(zinb_lice_a_brms_bayes) #0.316 
bayes_R2(zip_lice_a_brms_bayes) # zip 0.22
plot(zip_lice_a_brms_bayes)
plot(zinb_lice_a_brms_bayes)
mcmc_plot(zinb_lice_a_brms_bayes) # Dots represent means of posterior distribution along with 95% CrIs, as estimated by the bmod5 model
launch_shinystan()
pp_check(lice_a_brms_bayes, ndraws = 100)+ xlim(0, 5)  #  test for the model fit to the data .need to modify the scale of this plot posterior predictive checks, 100 random draws or distributions created by the model 
pp_check(lice_a_brms_bayes, type="bars", ndraws = 100)+ xlim(0, 20) 

pp_m<- brms::posterior_predict(zinb_lice_a_brms_bayes)
log1 <- scale_x_continuous(trans="log1p")
ppc_dens_overlay(y=lice_a_brms_bayes$data$total, pp_m[1:200, ])  + 
  coord_cartesian(xlim = c(0, 5))

pp_m<- brms::posterior_predict(zinb_lice_a_brms_bayes)
ppc_rootogram(y=zinb_lice_a_brms_bayes$data$total_lice, pp_m[1:200, ])  +   
  coord_cartesian(xlim = c(0, 5), ylim = c(0,30))

ppc_stat(y=lice_a_brms_bayes$data$total_lice, pp_m, stat ="prop_zero")


#Assumptions check model pglmm_ bayes
#remotes::install_github("Pakillo/DHARMa.helpers")

simulate_residuals <- dh_check_brms(zinb_lice_a_brms_bayes, integer = TRUE)
plot( simulate_residuals, form = ectos_birds_dff$sociality)
DHARMa::testDispersion(simulate_residuals)
DHARMa::testZeroInflation(simulate_residuals ) ## tests if there are more zeros in the data than expected from the simulations
testUniformity(simulate_residuals) #tests if the overall distribution conforms to expectations

# Plots BRMS 
color_scheme_set("teal") 
# model fit
png("data/data_analyses/models/model_plots/2.model_fit_ABUNDANCE_LICE_brms_phylo_multiple_obs_032123.png.png",width = 3000, height = 3000, res = 300, units = "px")
pp_m<- brms::posterior_predict(zinb_lice_a_brms_bayes)
ppc_rootogram(y=zinb_lice_a_brms_bayes$data$total_lice, pp_m[1:200, ])  +   
  coord_cartesian(xlim = c(0, 100), ylim = c(0,30))
dev.off()


# PLOT convergence and parameters distributions
png("data/data_analyses/models/model_plots/2.parameters_distribution_convergence_plot_model_ABUNDANCE_LICE_brms_phylo_multiple_obs_032123.png.png",width = 3000, height = 3000, res = 300, units = "px")
plot(zinb_lice_a_brms_bayes)
dev.off()

# plot posterior distributions 
color_scheme_set("teal")
brmstools::forest(ecto_p_brms_bayes, level = 0.95, show_data = TRUE)

estimates_plot<-mcmc_plot(zinb_lice_a_brms_bayes,prob=0.90, prob_outer=0.95,
                          variable = c("b_Intercept", "b_sociality1", "b_scaleelevation","sd_Powder.lvl__Intercept","sd_species__Intercept","sd_species_jetz__Intercept"),
                          type="areas") +
  labs(title="Posterior distributions Lice abundance", subtitle ="Lice abundance with medians and 95% intervals")+
  theme_minimal(20)+
  geom_vline(xintercept = 0, linetype = 2, colour = "grey40")+
  xlab("Estimate")

estimates_plot_intervals<-mcmc_plot(zinb_lice_a_brms_bayes,prob=0.90, prob_outer=0.95,point_est = "mean",
                                    variable = c("b_Intercept", "b_sociality1", "b_scaleelevation","sd_Powder.lvl__Intercept","sd_species__Intercept","sd_species_jetz__Intercept"),
                                    type="intervals") +
  labs(title="Posterior distributions Lice abundance", subtitle ="Lice abundance with medians and 95% intervals")+
  theme_minimal(20)+
  geom_vline(xintercept = 0, linetype = 2, colour = "grey40")+
  xlab("Estimate")

png("data/data_analyses/models/model_plots/2.parameters_intervals_plot_model_ABUNDANCE_LICE_brms_zinb_phylo_multiple_obs_032123.png",width = 3000, height = 3000, res = 300, units = "px")
estimates_plot_intervals
dev.off()
 
loo(zinb_lice_a_brms_bayes,zip_lice_a_brms_bayes,compare = TRUE)

# ###### 6.2 Model predictions PLOTS abundance lice----------------------------------

#descriptive plots 
ectos_birds_dff<-read.csv("data/data_manuscript/7.dff_all_ectos_prevalence_abundance_individual_elevation_FILE.csv", na.strings =c("","NA")) %>% 
  rename(elevation=elevation_extrapolated_date) %>%
  select(elevation, species_jetz, Powder.lvl,foraging_cat, sociality, total_lice) %>% 
  na.omit() %>% filter(species_jetz!="Premnoplex_brunnescens")  #Removing outliers for total mites

str(ectos_birds_dff)

#ectos_birds_dff <-ectos_birds_dff  %>% mutate_at("species_jetz", str_replace, " ", "_")
#ectos_birds_dff$elevation_cat<-as.factor(ectos_birds_dff$elevation_cat)
ectos_birds_dff$foraging_cat<-as.factor(ectos_birds_dff$foraging_cat)
ectos_birds_dff$species_jetz<-as.factor(ectos_birds_dff$species_jetz)
ectos_birds_dff$elevation<-as.numeric(ectos_birds_dff$elevation)
#ectos_birds_dff$elevation_midpoint<-as.numeric(ectos_birds_dff$elevation_midpoint)
ectos_birds_dff$sociality<-as.factor(ectos_birds_dff$sociality)
ectos_birds_dff$Powder.lvl<-as.factor(ectos_birds_dff$Powder.lvl)
ectos_birds_dff$total_lice<-as.numeric(ectos_birds_dff$total_lice)
ectos_birds_dff$species<- ectos_birds_dff$species_jetz # create a column for the species effect different to the phylogenetic one
names(ectos_birds_dff)
is.ultrametric(phylo)


# creating summaries per species

ectos_birds_dff_summary<-ectos_birds_dff %>% group_by(sociality) %>% 
  summarize( mean=mean(total_lice))

png("figures/figures_manuscript/Fig2_Lice_abundance_data_plot_sociality.png", width = 3000, height = 3000, res = 300, units = "px")
ggplot(data = ectos_birds_dff, aes(x = sociality, y=total_lice, color=species_jetz))+
  # geom_point(data = ectos_df, aes(x=sociality, y = proportion_ectoparasites),color="grey",size=2)+
  geom_jitter(data = ectos_birds_dff, aes(x=sociality, y =total_lice),color="darkgrey",size=3,width = 0.07, alpha=0.6)+
  geom_point(data = ectos_birds_dff_summary, aes(x=sociality, y =mean),color="cyan3",size=4)+
  scale_y_continuous("Lice abundance", limits = c(0,45)) +
  scale_x_discrete("Sociality", limits =c(0,1))+
  #geom_hline(yintercept = 0.775, linetype = "dashed")+
  theme_classic(40)
dev.off()


png("figures/figures_manuscript/Fig2_Lice_abundance_data_plot_elevation.png", width = 3000, height = 3000, res = 300, units = "px")
ggplot(data = ectos_birds_dff, aes(x =elevation, y=total_lice))+
  # geom_point(data = ectos_df, aes(x=sociality, y = proportion_ectoparasites),color="grey",size=2)+
  geom_jitter(data = ectos_birds_dff, aes(x=elevation, y =total_lice),color="darkgrey",size=3,width = 0.5, alpha=0.8)+
  scale_y_continuous("Lice abundance", limits = c(0,45)) +
  scale_x_continuous("Elevation", limits =c(0,4000))+
  #geom_smooth(method = "glm")+
  #geom_hline(yintercept = 0.775, linetype = "dashed")+
  theme_classic(40)
dev.off()

#Plotting posterior inference against the data I am not sure how to plot thi is the traformed scale


install.packages("bayestestR")
library(bayestestR)

predictions <- posterior_predict(zinb_lice_a_brms_bayes, newdata = ectos_birds_dff)

plot_data<- cbind(ectos_birds_dff, predictions)

ggplot(plot_data, aes(x = sociality, y = total_lice)) +
  geom_jitter() +
  xlab("Sociality") +
  ylab("Total Lice") +
  theme_classic() +
  geom_point(aes(y = predictions), color = "red")





ggplot(ectos_birds_dff, aes(x = sociality, y = total_lice)) +
  geom_jitter() +
  xlab("Sociality") +
  ylab("Total Lice") +
  theme_classic()

ggplot(ectos_birds_dff, aes(x = sociality, y = total_lice)) +
  geom_jitter() +
  xlab("Sociality") +
  ylab("Total Lice") +
  theme_classic() +
  geom_point(aes(y = predictions,x= sociality), color = "red")


in `geom_point()`:
  ! Problem while computing aesthetics.
ℹ Error occurred in the 2nd layer.
Caused by error in `check_aesthetics()`:
  ! Aesthetics must be either length 1 or the same as the data (756)
✖ Fix the following mappings: `y`








#https://bookdown.org/ajkurz/Statistical_Rethinking_recoded/linear-models.html

# Extracting posterior inference ( this is not working) 
posterior_summary(zinb_lice_a_brms_bayes)[1:8,]!!!!!

ectos_birds_dff %>%
  ggplot(aes(x=sociality, y = total_lice)) +
  geom_abline(intercept = fixef(zinb_lice_a_brms_bayes)[1], 
              slope     = fixef(zinb_lice_a_brms_bayes)[2]) +
  geom_point(shape = 1, size = 2, color = "royalblue") +
  theme_bw() +
  theme(panel.grid = element_blank())

fixef(zinb_lice_a_brms_bayes)

ectos_birds_dff %>%
  add_linpred_draws(zinb_lice_a_brms_bayes) %>%
  ggplot(aes(x = elevation, y =total_lice)) +
  stat_lineribbon(aes(y = .linpred), .width = c(.99, .95, .8, .5), color = "#08519C") +
  geom_point(data = ectos_birds_dff, size = 2) +
  scale_fill_brewer()

ectos_birds_dff  %>%
  # NOTE: this shows the use of ndraws to subsample within add_epred_draws()
  # ONLY do this IF you are planning to make spaghetti plots, etc.
  # NEVER subsample to a small sample to plot intervals, densities, etc.
  add_epred_draws(zinb_lice_a_brms_bayes, ndraws = 100) %>%   # sample 100 means from the posterior
  ggplot(aes(x = sociality, y = total_lice)) +
  geom_line(aes(y = .epred, group = .draw), alpha = 1/20, color = "#08519C") +
  geom_point(data = ectos_birds_dff) 





#some ideas here: https://cran.r-project.org/web/packages/tidybayes/vignettes/tidy-brms.html
# this is not workingunsure why
install.packages("modelr")
library(modelr)
epred_draws()

library(magrittr)
library(dplyr)
library(forcats)
library(tidyr)
library(modelr)
library(tidybayes)
library(ggplot2)
library(ggstance)
library(ggridges)
library(cowplot)
library(rstan)
library(brms)

# I am not sure how to do this 
# https://mjskay.github.io/tidybayes/reference/add_predicted_draws.html

ectos_birds_dff %>%
  group_by(sociality) %>%
  #data_grid(hp = seq_range(hp, n = 51)) %>%
  linpred_draws(object=zinb_lice_a_brms_bayes) %>% # plotting intervals it is always best to use all draws (omit ndraws)
  geom_line(aes(y =.linpred, group = paste(sociality, .draw)), alpha = 0.25, color="blue") +
  ggplot(aes(x = sociality, y = total_lice, color = sociality)) +
  #stat_lineribbon(aes(y =.prediction)) +
  geom_point(data =ectos_birds_dff) +
  scale_fill_brewer(palette = "Greys") +
  scale_color_brewer(palette = "Set2")


ectos_birds_dff %>%
  group_by(sociality) %>%
  #data_grid(hp = seq_range(hp, n = 51)) %>%
  predicted_draws(object=zinb_lice_a_brms_bayes) %>% # plotting intervals it is always best to use all draws (omit ndraws)
  ggplot(aes(x = sociality, y = total_lice, color = sociality)) +
  geom_line(aes(y =.prediction, group = paste(sociality, .draw)), alpha = 0.25, color="blue") +
  #stat_lineribbon(aes(y =.prediction)) +
  geom_point(data =ectos_birds_dff) +
  scale_fill_brewer(palette = "Greys") +
  scale_color_brewer(palette = "Set2")


#for [add_]epred_draws(), defaults to ".epred"

#for [add_]predicted_draws(), defaults to ".prediction".

#linpred_draws() ncludes a column with its name specified by the linpred argument (default ".linpred")

#add_linpred_draws() adds draws from the posterior linear predictors to the data. It corresponds to rstanarm::posterior_linpred() or brms::posterior_linpred(). Depending on the model type and additional parameters passed, this may be

add_predicted_draws(
  newdata,
  object,
  ...,
  value = ".prediction",
  ndraws = NULL,
  seed = NULL,
  re_formula = NULL,
  category = ".category",
  n
)

names(epred_draws)

# The oterh way of plotting this 

###_###_###
# PLOTS PREVALENCE [Presence Absence]
###_###_###

# PLotting the model predictions 
#### How well does PGLMM it predict the data
newdata <- data.frame(elevation = ectos_birds_dff$elevation,
                      species = ectos_birds_dff$species,
                      species_jetz = ectos_birds_dff$species_jetz,
                      sociality = ectos_birds_dff$sociality,
                      Powder.lvl=ectos_birds_dff$Powder.lvl)

predictions<- predict(zinb_lice_a_brms_bayes,newdata = newdata, type = "response" ) ##newdata = newdata
# or use fitedd instead
#predictions<- fitted(ecto_prevalence_pglmm,newdata = newdata) ##newdata = newdata

ectos_df_predicted <- cbind(ectos_birds_dff, predictions)

str(ectos_df_predicted)

# lets calculae the mean of the predited values on the untransformed scale
predictions_summary<- ectos_df_predicted %>% 
  group_by(sociality) %>%      
  dplyr::summarise(mean_prevalence= mean(Estimate), se=(Est.Error), n = n()) 

colnames(predictions_summary) <- c("sociality", "mean_prevalence", "se","n")

head(predictions_summary)
names(ectos_birds_dff)

# make a plot of model predictions (that also shows data)
png("figures/figures_manuscript/Fig1_Ectoparasite_prevalence_pglmm.png", width = 3000, height = 3000, res = 300, units = "px")
ggplot(data = ectos_birds_dff, aes(x = sociality, y = total_lice))+
  # geom_point(data = ectos_df, aes(x=sociality, y = proportion_ectoparasites),color="grey",size=2)+
  geom_jitter(data = ectos_df_predicted, aes(x=sociality, y = total_lice),color="grey",size=3,width = 0.07)+
  geom_segment(data = predictions_summary, aes(x = sociality, y = mean_prevalence, xend = sociality, yend =mean_prevalence+se, color="red"),show_guide = FALSE)+
  geom_segment(data = predictions_summary, aes(x = sociality, y = mean_prevalence, xend = sociality, yend =mean_prevalence-se, color="red"),show_guide = FALSE)+
  #geom_jitter(data = ectos_df_predicted, aes(x=sociality, y = Y_hat), color="red", size=4,shape=19,width = 0.07)+
  geom_point(data = predictions_summary, aes(x=sociality, y = mean_prevalence), color="red", size=4,shape=19)+
  scale_y_continuous("Lice abundance", limits = c(0,1)) +
  scale_x_discrete("Sociality")+
  #geom_hline(yintercept = 0.775, linetype = "dashed")+
  theme_classic(40)
dev.off()



ggplot(data = ectos_df_predicted, aes(x = sociality, y = proportion_ectoparasites))+
  # geom_point(data = ectos_df, aes(x=sociality, y = proportion_ectoparasites),color="grey",size=2)+
  geom_jitter(data = ectos_df_predicted, aes(x=sociality, y = proportion_ectoparasites),color="grey",size=3,width = 0.07)+
  geom_point(data = ectos_df_predicted, aes(x=sociality, y = proportion_ectoparasites), color="red", size=4,shape=19)+
  scale_y_continuous("Ectoparasites prevalence", limits = c(0,1)) +
  scale_x_discrete("Sociality")+
  geom_hline(yintercept = 0.775, linetype = "dashed")+
  theme_classic(40)


# ####### 6.Data processing abundance MITES----------------------------------------------------
ectos_birds_dff<-read.csv("data/data_manuscript/7.dff_all_ectos_prevalence_abundance_individual_elevation_FILE.csv", na.strings =c("","NA")) %>% 
  rename(elevation=elevation_extrapolated_date) %>%
  select(elevation, species_jetz, Powder.lvl,foraging_cat, sociality, total_mites, total_mesostigmatidae, total_no_feathers_mites) %>% 
  na.omit() %>% 
  filter(total_mites<300)  # removing outliers for total mites
#filter(total_mesostigmatidae<100)  # removing outliers for total mites
#filter(total_no_feathers_mites<100)  # removing outliers for total mites

# Outliers
assertr::insist(ectos_birds_dff, within_n_mads(50), total_mites)
assertr::insist(ectos_birds_dff, within_n_mads(2), total_no_feathers_mites)
View(ectos_birds_dff)

#ectos_birds_dff <-ectos_birds_dff  %>% mutate_at("species_jetz", str_replace, " ", "_")
#ectos_birds_dff$elevation_cat<-as.factor(ectos_birds_dff$elevation_cat)
ectos_birds_dff$foraging_cat<-as.factor(ectos_birds_dff$foraging_cat)
ectos_birds_dff$species_jetz<-as.factor(ectos_birds_dff$species_jetz)
ectos_birds_dff$elevation<-as.numeric(ectos_birds_dff$elevation)
#ectos_birds_dff$elevation_midpoint<-as.numeric(ectos_birds_dff$elevation_midpoint)
ectos_birds_dff$sociality<-as.factor(ectos_birds_dff$sociality)
ectos_birds_dff$Powder.lvl<-as.factor(ectos_birds_dff$Powder.lvl)
#ectos_birds_dff$total_lice<-as.factor(ectos_birds_dff$total_lice)
ectos_birds_dff$total_mites<-as.numeric(ectos_birds_dff$total_mites)
ectos_birds_dff$total_mesostigmatidae<-as.numeric(ectos_birds_dff$total_mesostigmatidae)
ectos_birds_dff$total_no_feathers_mites<-as.numeric(ectos_birds_dff$total_no_feathers_mites)
ectos_birds_dff$species<- ectos_birds_dff$species_jetz # create a column for the species effect different to the phylogenetic one

names(ectos_birds_dff)
is.ultrametric(phylo)

# Make sure the tips and the names on the file coincide and formating of name is consitent
phylo$edge.length  
phylo$tip.label
is.binary(phylo)

# Make sure this two are the same numbers 
a<-(as.data.frame(phylo$tip.label))%>% mutate(name=phylo$tip.label) %>% select(name) %>% arrange(desc(name))
b<-(as.data.frame(ectos_birds_dff$species_jetz)) %>% mutate(name=ectos_birds_dff$species_jetz) %>% select(name) %>% arrange(desc(name)) %>% distinct(name)

tip<-as.list(setdiff(a,b))
print(tip)

# Drop some tips USE IF NEED TO DROP SOME TIPS when using the full phylogeny
phylo<-drop.tip (phylo, tip$name) 

phy_cov<-ape::vcv(phylo, corr=TRUE)


# ####### 6.1.Analyses models abundance MITES--------------------------------------------------------

####
#NON_FEATHER MITES
####

ectos_birds_dff<-read.csv("data/data_manuscript/7.dff_all_ectos_prevalence_abundance_individual_elevation_FILE.csv", na.strings =c("","NA")) %>% 
  rename(elevation=elevation_extrapolated_date) %>%
  select(elevation, species_jetz, Powder.lvl,foraging_cat, sociality, total_no_feathers_mites) %>% 
  na.omit() %>% 
  filter(total_no_feathers_mites<100)  # removing outliers for total mites

# Removing premnoplex ( Outlier)
ectos_birds_dff<-read.csv("data/data_manuscript/7.dff_all_ectos_prevalence_abundance_individual_elevation_FILE.csv", na.strings =c("","NA")) %>% 
  rename(elevation=elevation_extrapolated_date) %>%
  select(elevation, species_jetz, Powder.lvl,foraging_cat, sociality, total_no_feathers_mites) %>% 
  na.omit() %>% 
  filter(species_jetz!="Premnoplex_brunnescens")  #Removing outliers for total mites


unique (ectos_birds_dff$species_jetz )
#filter(total_mesostigmatidae<100)  # removing outliers for total mites
#filter(total_no_feathers_mites<100)  # removing outliers for total mites

phylo<-read.nexus("data/phylo_data/consensus/1_consensus_birdtreeManu_ectos_prevalence.nex")  # This include speceis form manu and iquitos  so need to rpun the tree in the data processin section

# Outliers
assertr::insist(ectos_birds_dff, within_n_mads(50), total_mites)
assertr::insist(ectos_birds_dff, within_n_mads(2), total_no_feathers_mites)

#ectos_birds_dff <-ectos_birds_dff  %>% mutate_at("species_jetz", str_replace, " ", "_")
#ectos_birds_dff$elevation_cat<-as.factor(ectos_birds_dff$elevation_cat)
ectos_birds_dff$foraging_cat<-as.factor(ectos_birds_dff$foraging_cat)
ectos_birds_dff$species_jetz<-as.factor(ectos_birds_dff$species_jetz)
ectos_birds_dff$elevation<-as.numeric(ectos_birds_dff$elevation)
#ectos_birds_dff$elevation_midpoint<-as.numeric(ectos_birds_dff$elevation_midpoint)
ectos_birds_dff$sociality<-as.factor(ectos_birds_dff$sociality)
ectos_birds_dff$Powder.lvl<-as.factor(ectos_birds_dff$Powder.lvl)
#ectos_birds_dff$total_lice<-as.factor(ectos_birds_dff$total_lice)
#ectos_birds_dff$total_mites<-as.numeric(ectos_birds_dff$total_mites)
#ectos_birds_dff$total_mesostigmatidae<-as.numeric(ectos_birds_dff$total_mesostigmatidae)
ectos_birds_dff$total_no_feathers_mites<-as.numeric(ectos_birds_dff$total_no_feathers_mites)
ectos_birds_dff$species<- ectos_birds_dff$species_jetz # create a column for the species effect different to the phylogenetic one

names(ectos_birds_dff)
is.ultrametric(phylo)

# Make sure the tips and the names on the file coincide and formating of name is consitent
phylo$edge.length  
phylo$tip.label
is.binary(phylo)

# Make sure this two are the same numbers 
a<-(as.data.frame(phylo$tip.label))%>% mutate(name=phylo$tip.label) %>% select(name) %>% arrange(desc(name))
b<-(as.data.frame(ectos_birds_dff$species_jetz)) %>% mutate(name=ectos_birds_dff$species_jetz) %>% select(name) %>% arrange(desc(name)) %>% distinct(name)

tip<-as.list(setdiff(a,b))
print(tip)

# Drop some tips USE IF NEED TO DROP SOME TIPS when using the full phylogeny
phylo<-drop.tip (phylo, tip$name) 

phy_cov<-ape::vcv(phylo, corr=TRUE)

#a) model glmm
###_###_###
nf_mites_a_glmm <-glmer(total_no_feathers_mites~sociality+scale(elevation)+(1|Powder.lvl)+(1|species_jetz), #+elevation_midpoint+Powder.lvl
                     data = ectos_birds_dff, 
                     family = "poisson")
# Summary
summary(nf_mites_a_glmm )
rr2::R2(nf_mites_a_glmm)
fixef(nf_mites_a_glmm)
predict(nf_mites_a_glmm)
class(nf_mites_a_glmm) 

# Assumptions check
simulationOutput_nf_mites_a<- DHARMa::simulateResiduals(fittedModel =nf_mites_a_glmm, plot = F,integerResponse = T, re.form = NULL ) #quantreg=T
plot(simulationOutput_nf_mites_a)
testUniformity(simulationOutput_nf_mites_a) #tests if the overall distribution conforms to expectations
testOutliers(simulationOutput_nf_mites_a)#  tests if there are more simulation outliers than expected
testDispersion(simulationOutput_nf_mites_a) # tests if the simulated dispersion is equal to the observed dispersion
testQuantiles(simulationOutput_nf_mites_a) #fits a quantile regression or residuals against a predictor (default predicted value), and tests of this conforms to the expected quantile
testCategorical(simulationOutput_nf_mites_a, catPred = ectos_birds_dff$sociality)# tests residuals against a categorical predictor
testZeroInflation(simulationOutput_nf_mites_a) ## tests if there are more zeros in the data than expected from the simulations

# zero inflated?
#b) model pglmm
###_###_###

nf_mites_a_pglmm <-phyr::pglmm(total_no_feathers_mites~sociality+scale(elevation)+(1|species_jetz__)+(1|Powder.lvl),
                            data = ectos_birds_dff, 
                            family = "poisson",
                            cov_ranef = list(species_jetz= phylo), #class phylo
                            #bayes = TRUE,
                            add.obs.re = TRUE,
                            REML = TRUE, 
                            verbose = TRUE,
                            s2.init = .25) # what is this last parameter for

histogram(ectos_birds_dff$total_nf_mites) # id some outliers 

summary(nf_mites_a_pglmm)
rr2::R2(nf_mites_a_pglmm)
fixef(nf_mites_a_pglmm)
predict(ecto_a_pglmm)

# Assumptions check
simulationOutput_nf_mites_2<- DHARMa::simulateResiduals(fittedModel=nf_mites_a_pglmm, plot = F,integerResponse = T, re.form = NULL ) #quantreg=T
plot(simulationOutput_nf_mites_2)
testUniformity(simulationOutput_nf_mites_2) #tests if the overall distribution conforms to expectations
testOutliers(simulationOutput_nf_mites_2)#  tests if there are more simulation outliers than expected
testDispersion(simulationOutput_nf_mites_2) # tests if the simulated dispersion is equal to the observed dispersion
testQuantiles(simulationOutput_nf_mites_2) #fits a quantile regression or residuals against a predictor (default predicted value), and tests of this conforms to the expected quantile
testCategorical(simulationOutput_nf_mites_2, catPred = ectos_birds_dff$sociality)# tests residuals against a categorical predictor
testZeroInflation(simulationOutput_nf_mites_2) ## tests if there are more zeros in the data than expected from the simulations

# no outliers, but no homogenity of varianse KS and withig groups significant, and zero inflated

#c) model pglmm bayes: hierarchical Bayesian models fitted using integrated nested laplace approximation (INLA)
###_###_###

# zero.inflated poisson did not converge

nf_mites_a_pglmm_bayes <-phyr::pglmm(total_no_feathers_mites~sociality+scale(elevation)+(1|species_jetz__)+(1|Powder.lvl),
                                  data =ectos_birds_dff, 
                                  family ="zeroinflated.poisson", #POISSON  ="zeroinflated.poisson", #
                                  cov_ranef = list(species_jetz= phylo), #class phylo
                                  bayes = TRUE,
                                  verbose=FALSE,
                                  prior = "inla.default") # consider using    add.obs.re = T


#1) Summary of the model 
communityPGLMM.plot.re(x=nf_mites_a_pglmm_bayes) 
summary(nf_mites_a_pglmm_bayes )   
rr2::R2(nf_mites_a_pglmm_bayes)
class(ecto_abundance_pglmm_bayes)

launch_shinystan(ecto_abundance_pglmm_bayes)

# Plots

estimates_plot<-plot_bayes(nf_mites_a_pglmm_bayes) #  

#png("data/data_analyses/models/model_plots/2m.parameters_plot_model_nf_MITES_ABUNDANCE_zip_pglmm_phylo_multiple_obs_031623.png",width = 3000, height = 3000, res = 300, units = "px")
#estimates_plot
#dev.off()

# Assumptions check DHARMa does not work with pglm bayes=TRUE so I can not evaluate model fit.
simulationOutput_a_nf_mites_3<- DHARMa::simulateResiduals(fittedModel=nf_mites_a_pglmm_bayes, plot = F,integerResponse = T, re.form = NULL ) #quantreg=T
plot(simulationOutput_a_nf_mites_3)

#d) model pglmm bayes : zero_inflated_negbinomial, ind scaled elevation
###_###_###

zip_nf_mites_a_brms_bayes<-brm(total_no_feathers_mites~sociality+scale(elevation)+
                                  (1|gr(species_jetz, cov = phy_cov))+  
                                  (1|Powder.lvl) + 
                                  (1|species),
                                data=ectos_birds_dff,
                                family=zero_inflated_poisson,  #zero_inflated_negbinomial()
                                data2 = list(phy_cov=phy_cov),
                                iter=6000, warmup=3000,
                                thin=2,
                                control=list(adapt_delta=0.99, max_treedepth=12)) 
#saveRDS(zip_nf_mites_a_brms_bayes,"data/data_analyses/models/2m.model_ABUNDANCE_n_f_MITES_brms_zip_phylo_multiple_obs_wo_premnoplex.RDS")
zip_nf_mites_a_brms_bayes<-readRDS("data/data_analyses/models/2m.model_ABUNDANCE_n_f_MITES_brms_zip_phylo_multiple_obs_wo_premnoplex.RDS")
zip_nf_mites_a_brms_bayes<-readRDS("data/data_analyses/models/2m.model_ABUNDANCE_n_f_MITES_brms_zip_phylo_multiple_obs.RDS")

# removed Premnoplex_brunnescens


zinb_nf_mites_a_brms_bayes<-brm(total_no_feathers_mites~sociality+scale(elevation)+
                               (1|gr(species_jetz, cov = phy_cov))+  
                               (1|Powder.lvl) + 
                               (1|species),
                             data=ectos_birds_dff,
                             family=zero_inflated_negbinomial(),  #zero_inflated_negbinomial()
                             data2 = list(phy_cov=phy_cov),
                             iter=6000, warmup=3000,
                             thin=2,
                             control=list(adapt_delta=0.99, max_treedepth=12)) 

#saveRDS(zinb_nf_mites_a_brms_bayes, "data/data_analyses/models/2m.model_ABUNDANCE_n_f_MITES_brms_zinb_phylo_multiple_obs_wo_premnoplex.RDS")
zinb_nf_mites_a_brms_bayes<-readRDS("data/data_analyses/models/2m.model_ABUNDANCE_n_f_MITES_brms_zinb_phylo_multiple_obs_wo_premnoplex.RDS.RDS")
#zinb_nf_mites_a_brms_bayes<-readRDS("data/data_analyses/models/2m.model_ABUNDANCE_n_f_MITES_brms_zinb_phylo_multiple_obs.RDS")

# removed Premnoplex_brunnescens

zinb_nf_mites_a_brms_bayes_interactions<-brm(total_no_feathers_mites~sociality+scale(elevation)+sociality:scale(elevation)+
                                  (1|gr(species_jetz, cov = phy_cov))+  
                                  (1|Powder.lvl) + 
                                  (1|species),
                                data=ectos_birds_dff,
                                family=zero_inflated_negbinomial(),  #zero_inflated_negbinomial()
                                data2 = list(phy_cov=phy_cov),
                                iter=6000, warmup=3000,
                                thin=2,
                                control=list(adapt_delta=0.99, max_treedepth=12)) 


# Summarize the model
summary (zinb_nf_mites_a_brms_bayes_interactions)
fixef() # to get more detailed values for estimates
coef() # if you have group-level effects (hierarchical data)

# interpret the model 
#To test whether all regression coefficients are different from zero, we can look at the Credible Intervals that are listed in the summary output or we can visually represent them in density plots.
#If we do so, we clearly see that zero is not included in any of the density plots, meaning that we can be reasonably certain the regression coefficients are different from zero.
#INTERPRETATION:In the model, the parameter for Sociality means the expected difference between non_social(0) and social (1) with all other covariates held constant. we clearly see that zero is included in the density plot for sociality so there is not effect of sociality??
bayes_R2(zinb_nf_mites_a_brms_bayes) 
bayes_R2(zinb_nf_mites_a_brms_bayes_interactions)
plot(zinb_nf_mites_a_brms_bayes)
png("data/data_analyses/models/model_plots/Interactions.png",width = 3000, height = 3000, res = 300, units = "px")
mcmc_plot(zinb_nf_mites_a_brms_bayes_interactions,prob=0.90, prob_outer=0.95)# Dots represent means of posterior distribution along with 95% CrIs, as estimated by the bmod5 model
dev.off()
  launch_shinystan()
pp_check(zinb_nf_mites_a_brms_bayes, ndraws = 100)+ xlim(0, 5)  #  test for the model fit to the data .need to modify the scale of this plot posterior predictive checks, 100 random draws or distributions created by the model 
pp_check(zinb_nf_mites_a_brms_bayes, type="bars", ndraws = 100)+ xlim(0, 20) 

pp_m<- brms::posterior_predict(zinb_nf_mites_a_brms_bayes)
log1 <- scale_x_continuous(trans="log1p")
ppc_dens_overlay(y=zinb_nf_mites_a_brms_bayes$data$total_no_feathers_mites, pp_m[1:200, ])  + 
  coord_cartesian(xlim = c(0, 5))

ppc_rootogram(y=nf_mites_a_brms_bayes$data$total_no_feathers_mites, pp_m[1:200, ])  +   
  coord_cartesian(xlim = c(0, 100), ylim = c(0,30))

ppc_stat(y=nf_mites_a_brms_bayes$data$total_no_feathers_mites, pp_m, stat = "prop_zero")


#Assumptions check model pglmm_ bayes
#remotes::install_github("Pakillo/DHARMa.helpers")

simulate_residuals <- dh_check_brms(zinb_nf_mites_a_brms_bayes, integer = TRUE)
plot( simulate_residuals, form = ectos_birds_dff$sociality)
DHARMa::testDispersion(simulate_residuals)
DHARMa::testZeroInflation(simulate_residuals ) ## tests if there are more zeros in the data than expected from the simulations
testUniformity(simulate_residuals) #tests if the overall distribution conforms to expectations


# Plots BRMS 
# PLOTS EXCLUDING PREMNOPLEX  USE THIS MODEL
zinb_nf_mites_a_brms_bayes<-readRDS("data/data_analyses/models/2m.model_ABUNDANCE_n_f_MITES_brms_zinb_phylo_multiple_obs_wo_premnoplex.RDS.RDS")
color_scheme_set("green")

# PLots including premnoplex
zinb_nf_mites_a_brms_bayes<-readRDS("data/data_analyses/models/2m.model_ABUNDANCE_n_f_MITES_brms_zinb_phylo_multiple_obs.RDS")
color_scheme_set("orange")

# model convergence 
png("data/data_analyses/models/model_plots/2m.model_convergence_ABUNDANCE_nf_MITES_brms_zinb_phylo_multiple_obs_032123.png",width = 3000, height = 3000, res = 300, units = "px")
plot(zinb_nf_mites_a_brms_bayes)
dev.off()


# model fit
png("data/data_analyses/models/model_plots/2m.model_fit_ABUNDANCE_nf_MITES_brms_zinb_phylo_multiple_obs_032123.png",width = 3000, height = 3000, res = 300, units = "px")
pp_m<- brms::posterior_predict(zinb_nf_mites_a_brms_bayes)
ppc_rootogram(y=zinb_nf_mites_a_brms_bayes$data$total_no_feathers_mites, pp_m[1:200, ])  +   
  coord_cartesian(xlim = c(0, 100), ylim = c(0,30))
dev.off()

pp_m<- brms::posterior_predict(zinb_nf_mites_a_brms_bayes)

brmstools::forest(zinb_nf_mites_a_brms_bayes, level = 0.95, show_data = TRUE)

estimates_plot<-mcmc_plot(zinb_nf_mites_a_brms_bayes,prob=0.90, prob_outer=0.95,
                          variable = c("b_Intercept", "b_sociality1", "b_scaleelevation","sd_Powder.lvl__Intercept","sd_species__Intercept","sd_species_jetz__Intercept"),
                          type="areas") +
  labs(title="Posterior distribution mites (n_f)", subtitle ="with medians and 95% intervals")+
  theme_minimal(20)+
  geom_vline(xintercept = 0, linetype = 2, colour = "grey40")+
  xlab("Estimate")

estimates_plot_intervals<-mcmc_plot(zinb_nf_mites_a_brms_bayes,prob=0.90, prob_outer=0.95,point_est = "mean",
                                    variable = c("b_Intercept", "b_sociality1", "b_scaleelevation","sd_Powder.lvl__Intercept","sd_species__Intercept","sd_species_jetz__Intercept"),
                                    type="intervals") +
  labs(title="Posterior distribution mites(n_f)", subtitle ="with means and 95% intervals")+
  theme_minimal(20)+
  geom_vline(xintercept = 0, linetype = 2, colour = "grey40")+
  xlab("Estimate")

png("data/data_analyses/models/model_plots/2m.parameters_plot_model_intervals_ABUNDANCE_nf_MITES_brms_zinb_phylo_multiple_obs_032123.png",width = 3000, height = 3000, res = 300, units = "px")
estimates_plot_intervals
dev.off()

class(estimates_plot_intervals)
# Model comparisons 
loo(zip_nf_mites_a_brms_bayes,zinb_nf_mites_a_brms_bayes,compare = TRUE)

zip_nf_mites_a_brms_bayes





# ## ALL MITES ( excluded for now) -----------------------------------------------------------
# Mites
###_###_###
#a) model glmm
###_###_###
mites_a_glmm <-glmer(total_no_feathers_mites~sociality+scale(elevation)+(1|Powder.lvl)+(1|species_jetz), #+elevation_midpoint+Powder.lvl
                     data = ectos_birds_dff, 
                     family ="poisson")
# Summary
summary(mites_a_glmm )
rr2::R2(mites_a_glmm)
fixef(mites_a_glmm)
predict(mites_a_glmm)
class(mites_a_glmm) 

# Assumptions check
simulationOutput_a_mites<- DHARMa::simulateResiduals(fittedModel =mites_a_glmm, plot = F,integerResponse = T, re.form = NULL ) #quantreg=T
plot(simulationOutput_a_mites)
testUniformity(simulationOutput_a_mites) #tests if the overall distribution conforms to expectations
testOutliers(simulationOutput_a_mites)#  tests if there are more simulation outliers than expected
testDispersion(simulationOutput_a_mites) # tests if the simulated dispersion is equal to the observed dispersion
testQuantiles(simulationOutput_a_mites) #fits a quantile regression or residuals against a predictor (default predicted value), and tests of this conforms to the expected quantile
testCategorical(simulationOutput_a_mites, catPred = ectos_birds_dff$sociality)# tests residuals against a categorical predictor
testZeroInflation(simulationOutput_a_mites) ## tests if there are more zeros in the data than expected from the simulations

# zero inflated?
#b) model pglmm
###_###_###

mites_a_pglmm <-phyr::pglmm(total_mites~sociality+scale(elevation)+(1|species_jetz__)+(1|Powder.lvl),
                            data = ectos_birds_dff, 
                            family = "poisson",
                            cov_ranef = list(species_jetz= phylo), #class phylo
                            #bayes = TRUE,
                            add.obs.re = TRUE,
                            REML = TRUE, 
                            verbose = TRUE,
                            s2.init = .25) # what is this last parameter for

histogram(ectos_birds_dff$total_mites) # id some outliers 

summary(mites_a_pglmm)
rr2::R2(mites_a_pglmm)
fixef(mites_a_pglmm)
predict(ecto_a_pglmm)

# Assumptions check
simulationOutput_a_mites_2<- DHARMa::simulateResiduals(fittedModel=mites_a_pglmm, plot = F,integerResponse = T, re.form = NULL ) #quantreg=T
plot(simulationOutput_a_mites_2)
testUniformity(simulationOutput_a_mites_2) #tests if the overall distribution conforms to expectations
testOutliers(simulationOutput_a_mites_2)#  tests if there are more simulation outliers than expected
testDispersion(simulationOutput_a_mites_2) # tests if the simulated dispersion is equal to the observed dispersion
testQuantiles(simulationOutput_a_mites_2) #fits a quantile regression or residuals against a predictor (default predicted value), and tests of this conforms to the expected quantile
testCategorical(simulationOutput_a_mites_2, catPred = ectos_birds_dff$sociality)# tests residuals against a categorical predictor
testZeroInflation(simulationOutput_a_mites_2) ## tests if there are more zeros in the data than expected from the simulations

# no outliers, but no homogenity of varianse KS and withig groups significant, and zero inflated

#c) model pglmm bayes: hierarchical Bayesian models fitted using integrated nested laplace approximation (INLA)
###_###_###


zip_mites_a_pglmm_bayes <-phyr::pglmm(total_mites~sociality+scale(elevation)+(1|species_jetz__)+(1|Powder.lvl),
                                      data =ectos_birds_dff, 
                                      family ="zeroinflated.poisson", #POISSON  ="zeroinflated.poisson", #
                                      cov_ranef = list(species_jetz= phylo), #class phylo
                                      bayes = TRUE,
                                      verbose=FALSE,
                                      prior = "inla.default") # consider using    add.obs.re = T


#1) Summary of the model 
communityPGLMM.plot.re(x=zip_mites_a_pglmm_bayes) 
summary(zip_mites_a_pglmm_bayes)   
rr2::R2(zip_mites_a_pglmm_bayes)
class(ecto_abundance_pglmm_bayes)

launch_shinystan(ecto_abundance_pglmm_bayes)

# Plots

estimates_plot<-plot_bayes(zip_mites_a_pglmm_bayes ) #  

#png("data/data_analyses/models/model_plots/2m.parameters_plot_model_nf_MITES_ABUNDANCE_pglmm_phylo_multiple_obs_032123.png",width = 3000, height = 3000, res = 300, units = "px")
#estimates_plot
#dev.off()

# Assumptions check DHARMa does not work with pglm bayes=TRUE so I can not evaluate model fit.
simulationOutput_a_mites_3<- DHARMa::simulateResiduals(fittedModel=zip_mites_a_pglmm_bayes, plot = F,integerResponse = T, re.form = NULL ) #quantreg=T
plot(simulationOutput_a_mites_3)

#d) model pglmm bayes : zero_inflated_negbinomial, ind scaled elevation
###_###_###

zinb_mites_a_brms_bayes<-brm(total_mites~sociality+scale(elevation)+
                               (1|gr(species_jetz, cov = phy_cov))+  
                               (1|Powder.lvl) + 
                               (1|species),
                             data=ectos_birds_dff,
                             family=zero_inflated_negbinomial(),  #zero_inflated_negbinomial()
                             data2 = list(phy_cov=phy_cov),
                             iter=6000, warmup=3000,
                             thin=2,
                             control=list(adapt_delta=0.99, max_treedepth=12)) 

#saveRDS(zinb_mites_a_brms_bayes, "data/data_analyses/models/2m.model_ABUNDANCE_MITES_brms_zinb_phylo_multiple_obs.RDS")

zinb_mites_a_brms_bayes<-readRDS("data/data_analyses/models/2m.model_ABUNDANCE_MITES_brms_zinb_phylo_multiple_obs.RDS")

# Summarize the model
summary ()
fixef() # to get more detailed values for estimates
coef() # if you have group-level effects (hierarchical data)

# interpret the model 
#To test whether all regression coefficients are different from zero, we can look at the Credible Intervals that are listed in the summary output or we can visually represent them in density plots.
#If we do so, we clearly see that zero is not included in any of the density plots, meaning that we can be reasonably certain the regression coefficients are different from zero.
#INTERPRETATION:In the model, the parameter for Sociality means the expected difference between non_social(0) and social (1) with all other covariates held constant. we clearly see that zero is included in the density plot for sociality so there is not effect of sociality??
bayes_R2(zinb_mites_a_brms_bayes) 
plot(ecto_p_brms_bayes)
mcmc_plot(ecto_p_brms_bayes) # Dots represent means of posterior distribution along with 95% CrIs, as estimated by the bmod5 model
launch_shinystan()
pp_check(ecto_p_brms_bayes, ndraws = 100)+ xlim(0, 5)  #  test for the model fit to the data .need to modify the scale of this plot posterior predictive checks, 100 random draws or distributions created by the model 
pp_check(ecto_p_brms_bayes, type="bars", ndraws = 100)+ xlim(0, 20) 

pp_m<- brms::posterior_predict(ecto_p_brms_bayes)
log1 <- scale_x_continuous(trans="log1p")
ppc_dens_overlay(y=ecto_p_brms_bayes$data$ectoparasites_PA, pp_m[1:200, ])  + 
  coord_cartesian(xlim = c(0, 5))

ppc_rootogram(y=ecto_p_brms_bayes$data$ectoparasites_PA, pp_m[1:200, ])  +   
  coord_cartesian(xlim = c(0, 5), ylim = c(0,30))

ppc_stat(y=ecto_p_brms_bayes$data$ectoparasites_PA, pp_m, stat = "prop_zero")


#Assumptions check model pglmm_ bayes
#remotes::install_github("Pakillo/DHARMa.helpers")

simulate_residuals <- dh_check_brms(ecto_p_brms_bayes, integer = TRUE)
plot( simulate_residuals, form = ectos_birds_dff$sociality)
DHARMa::testDispersion(simulate_residuals)
DHARMa::testZeroInflation(simulate_residuals ) ## tests if there are more zeros in the data than expected from the simulations
testUniformity(simulate_residuals) #tests if the overall distribution conforms to expectations


# Plots BRMS 

#color_scheme_set("red")
brmstools::forest(ecto_p_brms_bayes, level = 0.95, show_data = TRUE)

estimates_plot<-mcmc_plot(lice_a_brms_bayes,prob=0.90, prob_outer=0.95,
                          variable = c("b_Intercept", "b_sociality1", "b_scaleelevation","sd_Powder.lvl__Intercept","sd_species__Intercept","sd_species_jetz__Intercept"),
                          type="areas") +
  labs(title="Posterior distributions", subtitle ="with medians and 95% intervals")+
  theme_minimal(20)+
  geom_vline(xintercept = 0, linetype = 2, colour = "grey40")+
  xlab("Estimate")

estimates_plot_intervals<-mcmc_plot(lice_a_brms_bayes,prob=0.90, prob_outer=0.95,point_est = "mean",
                                    variable = c("b_Intercept", "b_sociality1", "b_scaleelevation","sd_Powder.lvl__Intercept","sd_species__Intercept","sd_species_jetz__Intercept"),
                                    type="intervals") +
  labs(title="Posterior distributions", subtitle ="with means and 95% intervals")+
  theme_minimal(20)+
  geom_vline(xintercept = 0, linetype = 2, colour = "grey40")+
  xlab("Estimate")

png("data/data_analyses/models/model_plots/2.parameters_plot_model_LICE_ABUNDANCE_brms_zinb_phylo_multiple_obs_031623.png",width = 3000, height = 3000, res = 300, units = "px")
estimates_plot
dev.off()



#color_scheme_set("red")
brmstools::forest(ecto_p_brms_bayes, level = 0.95, show_data = TRUE)

estimates_plot<-mcmc_plot(lice_a_brms_bayes,prob=0.90, prob_outer=0.95,
                          variable = c("b_Intercept", "b_sociality1", "b_scaleelevation","sd_Powder.lvl__Intercept","sd_species__Intercept","sd_species_jetz__Intercept"),
                          type="areas") +
  labs(title="Posterior distributions", subtitle ="with medians and 95% intervals")+
  theme_minimal(20)+
  geom_vline(xintercept = 0, linetype = 2, colour = "grey40")+
  xlab("Estimate")

estimates_plot_intervals<-mcmc_plot(lice_a_brms_bayes,prob=0.90, prob_outer=0.95,point_est = "mean",
                                    variable = c("b_Intercept", "b_sociality1", "b_scaleelevation","sd_Powder.lvl__Intercept","sd_species__Intercept","sd_species_jetz__Intercept"),
                                    type="intervals") +
  labs(title="Posterior distributions", subtitle ="with means and 95% intervals")+
  theme_minimal(20)+
  geom_vline(xintercept = 0, linetype = 2, colour = "grey40")+
  xlab("Estimate")

png("data/data_analyses/models/model_plots/2.parameters_plot_model_LICE_ABUNDANCE_brms_zinb_phylo_multiple_obs_031623.png",width = 3000, height = 3000, res = 300, units = "px")
estimates_plot
dev.off()




# ###### 6.2 Model predictions plots abundance MITES----------------------------------
ectos_birds_dff<-read.csv("data/data_manuscript/7.dff_all_ectos_prevalence_abundance_individual_elevation_FILE.csv", na.strings =c("","NA")) %>% 
  rename(elevation=elevation_extrapolated_date) %>%
  select(elevation, species_jetz, Powder.lvl,foraging_cat, sociality, total_no_feathers_mites) %>% 
  na.omit() %>% filter(species_jetz!="Premnoplex_brunnescens")  #Removing outliers for total mites

str(ectos_birds_dff)

#ectos_birds_dff <-ectos_birds_dff  %>% mutate_at("species_jetz", str_replace, " ", "_")
#ectos_birds_dff$elevation_cat<-as.factor(ectos_birds_dff$elevation_cat)
ectos_birds_dff$foraging_cat<-as.factor(ectos_birds_dff$foraging_cat)
ectos_birds_dff$species_jetz<-as.factor(ectos_birds_dff$species_jetz)
ectos_birds_dff$elevation<-as.numeric(ectos_birds_dff$elevation)
#ectos_birds_dff$elevation_midpoint<-as.numeric(ectos_birds_dff$elevation_midpoint)
ectos_birds_dff$sociality<-as.factor(ectos_birds_dff$sociality)
ectos_birds_dff$Powder.lvl<-as.factor(ectos_birds_dff$Powder.lvl)
ectos_birds_dff$total_no_feathers_mites<-as.numeric(ectos_birds_dff$total_no_feathers_mites)
ectos_birds_dff$species<- ectos_birds_dff$species_jetz # create a column for the species effect different to the phylogenetic one
names(ectos_birds_dff)
is.ultrametric(phylo)


# creating summaries per species

ectos_birds_dff_summary<-ectos_birds_dff %>% group_by(sociality) %>% 
  summarize( mean=mean(total_no_feathers_mites))

png("figures/figures_manuscript/Fig2_Mites_abundance_data_plot_sociality.png", width = 3000, height = 3000, res = 300, units = "px")
ggplot(data = ectos_birds_dff, aes(x = sociality, y=total_no_feathers_mites))+
  # geom_point(data = ectos_df, aes(x=sociality, y = proportion_ectoparasites),color="grey",size=2)+
  geom_jitter(data = ectos_birds_dff, aes(x=sociality, y =total_no_feathers_mites),color="darkgrey",size=3,width =  0.5, alpha=0.8)+
  geom_point(data = ectos_birds_dff_summary, aes(x=sociality, y =mean),color="darkgreen",size=4)+
  scale_y_continuous("Mites abundance", limits = c(0,85)) +
  scale_x_discrete("Sociality", limits=c(0,1))+
  #geom_hline(yintercept = 0.775, linetype = "dashed")+
  theme_classic(40)
dev.off()


png("figures/figures_manuscript/Fig2_Mites_abundance_data_plot_elevation.png", width = 3000, height = 3000, res = 300, units = "px")
ggplot(data = ectos_birds_dff, aes(x =elevation, y=total_no_feathers_mites))+
  # geom_point(data = ectos_df, aes(x=sociality, y = proportion_ectoparasites),color="grey",size=2)+
  geom_jitter(data = ectos_birds_dff, aes(x=elevation, y =total_no_feathers_mites),color="darkgrey",size=3,width = 0.5, alpha=0.8)+
  scale_y_continuous("Mites abundance", limits = c(0,85)) +
  scale_x_continuous("Elevation", limits =c(0,4000))+
  #geom_smooth(method = "glm")+
  #geom_hline(yintercept = 0.775, linetype = "dashed")+
  theme_classic(40)
dev.off()

#Plotting posterior inference against the data I am not sure how to plot thi is the traformed scale
#https://bookdown.org/ajkurz/Statistical_Rethinking_recoded/linear-models.html
# ########7.Data processing diversity [In progress]--------------------------------------------------------
ectos_birds_dff_d<-read.csv("data/data_manuscript/", na.strings =c("","NA")) %>% 
  rename(elevation=elevation_extrapolated_date) %>%
  select(elevation, species_jetz, Powder.lvl,total_lice,foraging_cat, sociality, total_mites, total_mesostigmatidae, total_no_feathers_mites) %>% 
  na.omit() %>% 
  filter(total_mites<300) 
  
assertr::insist(ectos_birds_dff, within_n_mads(50), total_mites)
assertr::insist(ectos_birds_dff, within_n_mads(2), total_no_feathers_mites)


View(ectos_birds_dff)

#ectos_birds_dff <-ectos_birds_dff  %>% mutate_at("species_jetz", str_replace, " ", "_")
#ectos_birds_dff$elevation_cat<-as.factor(ectos_birds_dff$elevation_cat)
ectos_birds_dff$foraging_cat<-as.factor(ectos_birds_dff$foraging_cat)
ectos_birds_dff$species_jetz<-as.factor(ectos_birds_dff$species_jetz)
ectos_birds_dff$elevation<-as.numeric(ectos_birds_dff$elevation)
#ectos_birds_dff$elevation_midpoint<-as.numeric(ectos_birds_dff$elevation_midpoint)
ectos_birds_dff$sociality<-as.factor(ectos_birds_dff$sociality)
ectos_birds_dff$Powder.lvl<-as.factor(ectos_birds_dff$Powder.lvl)
ectos_birds_dff$total_lice<-as.factor(ectos_birds_dff$total_lice)
ectos_birds_dff$total_mites<-as.factor(ectos_birds_dff$total_mites)
ectos_birds_dff$total_mesostigmatidae<-as.factor(ectos_birds_dff$total_mesostigmatidae)
ectos_birds_dff$total_no_feathers_mites<-as.factor(ectos_birds_dff$total_no_feathers_mites)

names(ectos_birds_dff)
is.ultrametric(phylo)

# Make sure the tips and the names on the file coincide and formating of name is consitent
phylo$edge.length  
phylo$tip.label
is.binary(phylo)

# Make sure this two are the same numbers 
a<-(as.data.frame(phylo$tip.label))%>% mutate(name=phylo$tip.label) %>% select(name) %>% arrange(desc(name))
b<-(as.data.frame(ectos_dff$species_jetz)) %>% mutate(name=ectos_dff$species_jetz) %>% select(name) %>% arrange(desc(name)) %>% distinct(name)

tip<-as.list(setdiff(a,b))
print(tip)

# Drop some tips USE IF NEED TO DROP SOME TIPS when using the full phylogeny
phylo<-drop.tip (phylogeny_prevalence, tip$name)

# ####### 7.1 Analyses models diversity [In progress]--------------------------------------------------------


###_###_###
#d) model pglmm bayes
###_###_###

# Summarize the model
# Summarize the model
summary ()
fixef() # to get more detailed values for estimates
coef() # if you have group-level effects (hierarchical data)

# interpret the model 
#To test whether all regression coefficients are different from zero, we can look at the Credible Intervals that are listed in the summary output or we can visually represent them in density plots.
#If we do so, we clearly see that zero is not included in any of the density plots, meaning that we can be reasonably certain the regression coefficients are different from zero.
#INTERPRETATION:In the model, the parameter for Sociality means the expected difference between non_social(0) and social (1) with all other covariates held constant. we clearly see that zero is included in the density plot for sociality so there is not effect of sociality??
bayes_R2(ecto_p_brms_bayes) 
plot(ecto_p_brms_bayes)
mcmc_plot(ecto_p_brms_bayes) # Dots represent means of posterior distribution along with 95% CrIs, as estimated by the bmod5 model
launch_shinystan()
pp_check(ecto_p_brms_bayes, ndraws = 100)+ xlim(0, 5)  #  test for the model fit to the data .need to modify the scale of this plot posterior predictive checks, 100 random draws or distributions created by the model 
pp_check(ecto_p_brms_bayes, type="bars", ndraws = 100)+ xlim(0, 20) 

pp_m<- brms::posterior_predict(ecto_p_brms_bayes)
log1 <- scale_x_continuous(trans="log1p")
ppc_dens_overlay(y=ecto_p_brms_bayes$data$ectoparasites_PA, pp_m[1:200, ])  + 
  coord_cartesian(xlim = c(0, 5))

ppc_rootogram(y=ecto_p_brms_bayes$data$ectoparasites_PA, pp_m[1:200, ])  +   
  coord_cartesian(xlim = c(0, 5), ylim = c(0,30))

ppc_stat(y=ecto_p_brms_bayes$data$ectoparasites_PA, pp_m, stat = "prop_zero")


#Assumptions check model pglmm_ bayes
#remotes::install_github("Pakillo/DHARMa.helpers")

simulate_residuals <- dh_check_brms(ecto_p_brms_bayes, integer = TRUE)
plot( simulate_residuals, form = ectos_birds_dff$sociality)
DHARMa::testDispersion(simulate_residuals)
DHARMa::testZeroInflation(simulate_residuals ) ## tests if there are more zeros in the data than expected from the simulations
testUniformity(simulate_residuals) #tests if the overall distribution conforms to expectations

# Plots BRMS 

color_scheme_set("red")
brmstools::forest(ecto_p_brms_bayes, level = 0.95, show_data = TRUE)

mcmc_plot(ecto_p_brms_bayes ,prob=0.90, prob_outer=0.95) +
  ggtitle("Posterior intervals")+
  theme_minimal(15)+
  geom_vline(xintercept = 0, linetype = 2, colour = "grey40")+
  xlab("Estimate")

posterior <- as.array(ecto_p_brms_bayes)
mcmc_areas(posterior  ,prob=0.90, prob_outer=0.95)



# 0.N Libraries networks --------------------------------------------------------------


# # 1.N Data import-----------------------------------------------------------------

dff_ectos_network_individual_metrics<-read.csv("data/data_manuscript/7.dff_all_ectos_network_metrics_individuals_FILE.csv",na.strings =c("","NA"))%>% 
  select(elevation_extrapolated_date, species_jetz, Powder.lvl,foraging_cat, sociality, ectoparasites_PA,total_lice,total_mites, total_no_feathers_mites,total_mesostigmatidae,degree, w_degree) %>% 
  rename(elevation=elevation_extrapolated_date) %>%
  na.omit() %>% filter(species_jetz!="Premnoplex_brunnescens")  #Removing outliers for total mites

names(df_ectos_network_individual_metrics)

unique(df_ectos_network_individual_metrics$species_jetz) # this is teh total species that are in flocks taht we have samples for

phylo<-read.nexus("data/phylo_data/consensus/1_consensus_birdtreeManu_ectos_prevalence.nex")  # This include speceis form manu and iquitos social and non social so we need to trim it 


# ####### 8.Data processing networks lice--------------------------------------------------------

# Occurrence 

# Lice

dff_ectos_network_individual_metrics<-read.csv("data/data_manuscript/7.dff_all_ectos_network_metrics_individuals_FILE.csv",na.strings =c("","NA"))%>% 
  select(elevation_extrapolated_date, species_jetz, Powder.lvl,foraging_cat, sociality,total_lice, degree, w_degree) %>% 
  rename(elevation=elevation_extrapolated_date) %>%
  na.omit() %>% filter(species_jetz!="Premnoplex_brunnescens")  #Removing outliers for total mites

unique(dff_ectos_network_individual_metrics$species_jetz) # this is teh total species that are in flocks taht we have samples for

phylo<-read.nexus("data/phylo_data/consensus/1_consensus_birdtreeManu_ectos_prevalence.nex")  # This include speceis form manu and iquitos social and non social so we need to trim it 

# Make sure the tips and the names on the file coincide and formating of name is consitent
phylo$edge.length  
phylo$tip.label
is.binary(phylo)

# Make sure this two are the same numbers 
a<-(as.data.frame(phylo$tip.label))%>% mutate(name=phylo$tip.label) %>% select(name) %>% arrange(desc(name))
b<-(as.data.frame(dff_ectos_network_individual_metrics$species_jetz)) %>% mutate(name=dff_ectos_network_individual_metrics$species_jetz) %>% select(name) %>% arrange(desc(name)) %>% distinct(name)

tip<-as.list(setdiff(a,b))
print(tip)

# Drop some tips USE IF NEED TO DROP SOME TIPS when using the full phylogeny
phylo<-drop.tip (phylo, tip$name) 

# phylogenetic correlation structure, create a covariance matrix of species
phy_cov<-ape::vcv(phylo, corr=TRUE)

# data structure 

#dff_ectos_network_individual_metrics$elevation_cat<-as.factor(ectos_birds_dff$elevation_cat)
dff_ectos_network_individual_metrics$foraging_cat<-as.factor(dff_ectos_network_individual_metrics$foraging_cat)
dff_ectos_network_individual_metrics$species_jetz<-as.factor(dff_ectos_network_individual_metrics$species_jetz)
dff_ectos_network_individual_metrics$elevation<-as.numeric(dff_ectos_network_individual_metrics$elevation)
dff_ectos_network_individual_metrics$degree<-as.numeric(dff_ectos_network_individual_metrics$degree)
dff_ectos_network_individual_metrics$w_degree<-as.numeric(dff_ectos_network_individual_metrics$w_degree)

#ectos_birds_dff$elevation_midpoint<-as.numeric(ectos_birds_dff$elevation_midpoint)
dff_ectos_network_individual_metrics$sociality<-as.factor(dff_ectos_network_individual_metrics$sociality)
dff_ectos_network_individual_metrics$Powder.lvl<-as.factor(dff_ectos_network_individual_metrics$Powder.lvl)
dff_ectos_network_individual_metrics$total_lice<-as.numeric(dff_ectos_network_individual_metrics$total_lice)
dff_ectos_network_individual_metrics$species<-dff_ectos_network_individual_metrics$species_jetz # create a column for the species effect different to the phylogenetic one
names(dff_ectos_network_individual_metrics)
is.ultrametric(phylo)
 LICE 


# ####### 8.1. Analyses models networks Lice --------------------------------------------------------

# Lice
#Degree!!!
###_###_###
#a) model glmm
###_###_###
str(dff_ectos_network_individual_metrics)
lice_a_glmm_degree <-glmer(total_lice~degree+scale(elevation)+(1|Powder.lvl)+(1|species_jetz), #+elevation_midpoint+Powder.lvl
                    data = dff_ectos_network_individual_metrics, 
                    family = "poisson")

# Summary
summary(lice_a_glmm_degree )
rr2::R2(lice_a_glmm_degree)# R2 Predicted 0.38
fixef(lice_a_glmm_degree)
predict(lice_a_glmm_degree)
class(lice_a_glmm_degree) 

# Assumptions check
simulationOutput_a_lice<- DHARMa::simulateResiduals(fittedModel =lice_a_glmm_degree, plot = F,integerResponse = T, re.form = NULL ) #quantreg=T
plot(simulationOutput_a_lice)
testUniformity(simulationOutput_a_lice) #tests if the overall distribution conforms to expectations
testOutliers(simulationOutput_a_lice)#  tests if there are more simulation outliers than expected
testDispersion(simulationOutput_a_lice) # tests if the simulated dispersion is equal to the observed dispersion
testQuantiles(simulationOutput_a_lice) #fits a quantile regression or residuals against a predictor (default predicted value), and tests of this conforms to the expected quantile
testCategorical(simulationOutput_a_lice, catPred = dff_ectos_network_individual_metrics$sociality)# tests residuals against a categorical predictor
testZeroInflation(simulationOutput_a_lice) ## tests if there are more zeros in the data than expected from the simulations

# zero inflated, overdispersed, with outliers
#b) model pglmm
###_###_###

lice_a_pglmm_degree<-phyr::pglmm(total_lice~degree+scale(elevation)+(1|species_jetz__)+(1|Powder.lvl),
                           data = dff_ectos_network_individual_metrics, 
                           family = "poisson",
                           cov_ranef = list(species_jetz= phylo), #class phylo
                           #bayes = TRUE,
                           add.obs.re = TRUE,
                           REML = TRUE, 
                           verbose = TRUE,
                           s2.init = .25) # what is this last parameter for

histogram(dff_ectos_network_individual_metrics$total_lice) # id some outliers 
plot_data(lice_a_pglmm_degree)
summary(lice_a_pglmm_degree)
rr2::R2(lice_a_pglmm_degree)
fixef(lice_a_pglmm_degree)
predict(lice_a_pglmm_degree)

# Assumptions check
simulationOutput_a_lice_2<- DHARMa::simulateResiduals(fittedModel=lice_a_pglmm_degree, plot = F,integerResponse = T, re.form = NULL ) #quantreg=T
plot(simulationOutput_a_lice_2)
testUniformity(simulationOutput_a_lice_2) #tests if the overall distribution conforms to expectations
testOutliers(simulationOutput_a_lice_2)#  tests if there are more simulation outliers than expected
testDispersion(simulationOutput_a_lice_2) # tests if the simulated dispersion is equal to the observed dispersion
testQuantiles(simulationOutput_a_lice_2) #fits a quantile regression or residuals against a predictor (default predicted value), and tests of this conforms to the expected quantile
#testCategorical(simulationOutput_a_lice_2, catPred = dff_ectos_network_individual_metrics$sociality)# tests residuals against a categorical predictor
testZeroInflation(simulationOutput_a_lice_2) ## tests if there are more zeros in the data than expected from the simulations

# no outliers, but no homogenity of varianse KS and withig groups significant, and zero inflated

#c) model pglmm bayes: hierarchical Bayesian models fitted using integrated nested laplace approximation (INLA)
###_###_###

zip_lice_a_pglmm_bayes_degree<-phyr::pglmm(total_lice~scale(degree, center=TRUE)+scale(elevation)+(1|species_jetz__)+(1|Powder.lvl),
                                     data =dff_ectos_network_individual_metrics, 
                                     family ="zeroinflated.poisson", #POISSON  ="zeroinflated.poisson", #
                                     cov_ranef = list(species_jetz= phylo), #class phylo
                                     bayes = TRUE,
                                     verbose=FALSE,
                                     prior = "inla.default",
                                     control.inla=list(control.vb=list(emergency=30))) # consider using    add.obs.re = T
# control = list(adapt_delta = 0.99) 
#saveRDS(zip_lice_a_pglmm_bayes_degree , "data/data_analyses/models/2.model_ABUNDANCE_LICE_pglmm_zip_phylo_multiple_obs_17032023_degree.RDS")
zip_lice_a_pglmm_bayes<-readRDS ("data/data_analyses/models/2.model_ABUNDANCE_LICE_pglmm_zip_phylo_multiple_obs_17032023_degree.RDS")
#1) Summary of the model 
summary(zip_lice_a_pglmm_bayes_degree)   
rr2::R2(zip_lice_a_pglmm_bayes_degree)
class(zip_lice_a_pglmm_bayes_degree)

launch_shinystan(zip_lice_a_pglmm_bayes_degree)

# Plots
estimates_plot<-plot_bayes(zip_lice_a_pglmm_bayes_degree ) # for some reason does not allow me to plot zero inflated poisson 
#png("data/data_analyses/models/model_plots/2.parameters_plot_model_LICE_ABUNDANCE_pglmm_zip_phylo_multiple_obs_032123.png",width = 3000, height = 3000, res = 300, units = "px")
#estimates_plot
#dev.off()

# Assumptions check DHARMa does not work with pglm bayes=TRUE so I can not evaluate model fit.
simulationOutput_a_lice_3<- DHARMa::simulateResiduals(fittedModel=zip_lice_a_pglmm_bayes_degree, plot = F,integerResponse = T, re.form = NULL ) #quantreg=T
plot(simulationOutput_a_lice_3)
testUniformity(simulationOutput_a_lice_3) #tests if the overall distribution conforms to expectations
testOutliers(simulationOutput_a_lice_3)#  tests if there are more simulation outliers than expected
testDispersion(simulationOutput_a_lice_3) # tests if the simulated dispersion is equal to the observed dispersion
testQuantiles(simulationOutput_a_lice_3) #fits a quantile regression or residuals against a predictor (default predicted value), and tests of this conforms to the expected quantile
testCategorical(simulationOutput_a_lice_3, catPred = dff_ectos_network_individual_metrics$sociality)# tests residuals against a categorical predictor
testZeroInflation(simulationOutput_a_lice_3) ## tests if there are more zeros in the data than expected from the simulations


#d) model brms bayes : zero_inflated_negbinomial, ind scaled elevation
###_###_###

improved_prior <- prior(normal(0, 10), class = "b") + 
  prior(student_t(3, 0, 10), class = "Intercept") + 
  prior(normal(0, 10), class = "sd") 

improved_prior <- prior(normal(0, 10), class = "b") + 
  prior(student_t(3, 0, 10), class = "Intercept") + 
  prior(normal(0, 10), class = "sd") 

non_informative_prior<- c(
  prior(normal(0, 10), class = "Intercept"),
  prior(normal(0, 10), class = "b", coef = c("scale(w_degree, center=TRUE)", "scale(elevation)")),
  prior(cauchy(0, 2.5), class = "sd", group = c("species_jetz", "species", "Powder.lvl"), nlpar = "b"),
  prior(uniform(0, Inf), class = "b", nlpar = "zi")
)

# degree zinb

zinb_lice_a_brms_bayes_degree<-brm(total_lice~scale(degree, center=TRUE) +scale(elevation)+
                              (1|gr(species_jetz, cov = phy_cov))+  
                              (1|Powder.lvl) + 
                              (1|species),
                            data=dff_ectos_network_individual_metrics,
                            family=zero_inflated_negbinomial(),  #zero_inflated_negbinomial()
                            data2 = list(phy_cov=phy_cov),
                            iter=8000, warmup=4000,thin=4,
                            control=list(adapt_delta=0.99, max_treedepth=12)) 
#                            prior = improved_prior,


#saveRDS(zinb_lice_a_brms_bayes_degree, "data/data_analyses/models/2.model_ABUNDANCE_LICE_brms_zinb_phylo_multiple_obs_25032023_degree.RDS")
zinb_lice_a_brms_bayes_degree<-readRDS("data/data_analyses/models/2.model_ABUNDANCE_LICE_brms_zinb_phylo_multiple_obs_25032023_degree.RDS")

hypothesis(zinb_lice_a_brms_bayes_degree,"scale(degree)>0", alpha=0.05 ) # increase of degree increases lice abundance
hypothesis(zinb_lice_a_brms_bayes_degree,"total_lice=degree+species", alpha=0.05 ) # increase of degree increases lice abundance
###



#saveRDS(zinb_lice_a_brms_bayes_degree, "data/data_analyses/models/2.model_ABUNDANCE_LICE_brms_zinb_phylo_multiple_obs_25032023_degree.RDS")
zinb_lice_a_brms_bayes_degree<-readRDS("data/data_analyses/models/2.model_ABUNDANCE_LICE_brms_zinb_phylo_multiple_obs_25032023_degree.RDS")


#  exploring zero inflated poisson 
# degree zip

zip_lice_a_brms_bayes_degree<-brm(total_lice~scale(degree, center=TRUE) +scale(elevation)+
                             (1|gr(species_jetz, cov = phy_cov))+  
                             (1|species),
                           data=dff_ectos_network_individual_metrics,
                           family=zero_inflated_poisson(),  #zero_inflated_negbinomial()
                           data2 = list(phy_cov=phy_cov),
                           iter=8000, warmup=4000,
                           thin=4,
                           control=list(adapt_delta=0.99, max_treedepth=12)) 

zip_lice_a_brms_bayes_degree
#saveRDS(zip_lice_a_brms_bayes_degree, "data/data_analyses/models/2.model_ABUNDANCE_LICE_brms_zip_phylo_multiple_obs_21032023_degree.RDS")

# some outliers and smaller R2, overall performed porrly compared to zeroinflated negative binomial)_


# Lice
#w_Degree!!!

zinb_lice_a_brms_bayes_w_degree<-brm(total_lice~scale(w_degree, center=TRUE) +scale(elevation)+
                                     (1|gr(species_jetz, cov = phy_cov))+  
                                     (1|Powder.lvl) + 
                                     (1|species),
                                   data=dff_ectos_network_individual_metrics,
                                   family=zero_inflated_negbinomial(),  #zero_inflated_negbinomial()
                                   data2 = list(phy_cov=phy_cov),
                                   iter=8000, warmup=4000,thin=4,
                                   #prior=improved_prior,
                                   control=list(adapt_delta=0.99, max_treedepth=12)) 


saveRDS(zinb_lice_a_brms_bayes_w_degree, "data/data_analyses/models/2.model_ABUNDANCE_LICE_brms_zinb_phylo_multiple_obs_25032023_w_degree.RDS")
zinb_lice_a_brms_bayes_w_degree<-readRDS("data/data_analyses/models/2.model_ABUNDANCE_LICE_brms_zinb_phylo_multiple_obs_25032023_w_degree.RDS")

hypothesis(zinb_lice_a_brms_bayes_degree,"degree>0", alpha=0.05 ) # increase of degree increases lice abundance
hypothesis(zinb_lice_a_brms_bayes_degree,"total_lice=degree+species", alpha=0.05 ) # increase of degree increases lice abundance

#  exploring zero inflated poisson 

zip_lice_a_brms_bayes_w_degree<-brm (total_lice~scale(w_degree,center=TRUE)+scale(elevation)+
                                    (1|gr(species_jetz, cov = phy_cov))+  
                                    (1|species),
                                  data=dff_ectos_network_individual_metrics,
                                  family=zero_inflated_poisson(),  #zero_inflated_negbinomial()
                                  data2 = list(phy_cov=phy_cov),
                                  iter=8000, warmup=4000,
                                  thin=2,
                                  control=list(adapt_delta=0.99, max_treedepth=12)) 

zip_lice_a_brms_bayes_degree
#saveRDS(zip_lice_a_brms_bayes_w_degree, "data/data_analyses/models/2.model_ABUNDANCE_LICE_brms_zip_phylo_multiple_obs_21032023_w_degree.RDS")


# Summarize the models
summary(zinb_lice_a_brms_bayes_degree)
fixef(zinb_lice_a_brms_bayes_degree) # to get more detailed values for estimates
coef(zinb_lice_a_brms_bayes_degree) # if you have group-level effects (hierarchical data)

# interpret the model 
#To test whether all regression coefficients are different from zero, we can look at the Credible Intervals that are listed in the summary output or we can visually represent them in density plots.
#If we do so, we clearly see that zero is not included in any of the density plots, meaning that we can be reasonably certain the regression coefficients are different from zero.
#INTERPRETATION:In the model, the parameter for Sociality means the expected difference between non_social(0) and social (1) with all other covariates held constant. we clearly see that zero is included in the density plot for sociality so there is not effect of sociality??
bayes_R2(zip_lice_a_brms_bayes_w_degree) # zing 0.31
plot(zinb_lice_a_brms_bayes_degree)
mcmc_plot(zip_lice_a_brms_bayes_w_degree) # Dots represent means of posterior distribution along with 95% CrIs, as estimated by the bmod5 model
launch_shinystan()
pp_check(zinb_lice_a_brms_bayes_degree, ndraws = 100)+ xlim(0, 5)  #  test for the model fit to the data .need to modify the scale of this plot posterior predictive checks, 100 random draws or distributions created by the model 
pp_check(zinb_lice_a_brms_bayes_degree, type="bars", ndraws = 100)+ xlim(0, 20) 

pp_m<- brms::posterior_predict(zinb_lice_a_brms_bayes_degree)
log1 <- scale_x_continuous(trans="log1p")
ppc_dens_overlay(y=zinb_lice_a_brms_bayes_degree$data$total, pp_m[1:200, ])  + 
  coord_cartesian(xlim = c(0, 5))

pp_m<- brms::posterior_predict(zinb_lice_a_brms_bayes_degree)
ppc_rootogram(y=zinb_lice_a_brms_bayes_degree$data$total_lice, pp_m[1:200, ])  +   
  coord_cartesian(xlim = c(0, 5), ylim = c(0,30))

ppc_stat(y=zinb_lice_a_brms_bayes_degree$data$total_lice, pp_m, stat ="prop_zero")


#Assumptions check model pglmm_ bayes
#remotes::install_github("Pakillo/DHARMa.helpers")

simulate_residuals <- dh_check_brms(zinb_lice_a_brms_bayes_degree, integer = TRUE)
plot( simulate_residuals, form = dff_ectos_network_individual_metrics$sociality)
DHARMa::testDispersion(simulate_residuals)
DHARMa::testZeroInflation(simulate_residuals ) ## tests if there are more zeros in the data than expected from the simulations
testUniformity(simulate_residuals) #tests if the overall distribution conforms to expectations

###_###_###_###_
# Plots BRMS_degree
zinb_lice_a_brms_bayes_degree<-readRDS("data/data_analyses/models/2.model_ABUNDANCE_LICE_brms_zinb_phylo_multiple_obs_25032023_degree.RDS")
color_scheme_set("yellow") 
# model convergence 
png("data/data_analyses/models/model_plots/4l.model_convergence_ABUNDANCE_LICE_brms_zinb_phylo_multiple_obs_032123_DEGREE.png",width = 3000, height = 3000, res = 300, units = "px")
plot(zinb_lice_a_brms_bayes_degree)
dev.off()

# model fit
png("data/data_analyses/models/model_plots/4l.model_fit_ABUNDANCE_LICE_brms_zinb_phylo_multiple_obs_032123_DEGREE.png",width = 3000, height = 3000, res = 300, units = "px")
pp_m<- brms::posterior_predict(zinb_lice_a_brms_bayes_degree)
ppc_rootogram(y=zinb_lice_a_brms_bayes_degree$data$total_lice, pp_m[1:200, ])  +   
  coord_cartesian(xlim = c(0, 100), ylim = c(0,30))
dev.off()

# FOREST PLOTS OF POSTEIOR INTERVALS

estimates_plot<-mcmc_plot(zinb_lice_a_brms_bayes_degree,prob=0.90, prob_outer=0.95,
                          variable = c("b_Intercept", "b_scaledegreecenterEQTRUE", "b_scaleelevation","sd_Powder.lvl__Intercept","sd_species__Intercept","sd_species_jetz__Intercept"),
                          type="areas") +
  labs(title="Posterior distribution Lice [NETWORK]", subtitle ="with medians and 95% intervals")+
  theme_minimal(20)+
  geom_vline(xintercept = 0, linetype = 2, colour = "grey40")+
  xlab("Estimate")

estimates_plot_intervals<-mcmc_plot(zinb_lice_a_brms_bayes_degree,prob=0.90, prob_outer=0.95,point_est = "mean",
                                    variable = c("b_Intercept", "b_scaledegreecenterEQTRUE", "b_scaleelevation","sd_Powder.lvl__Intercept","sd_species__Intercept","sd_species_jetz__Intercept"),
                                    type="intervals") +
  labs(title="Posterior distribution Lice [NETWORK]", subtitle ="with means and 95% intervals")+
  theme_minimal(20)+
  geom_vline(xintercept = 0, linetype = 2, colour = "grey40")+
  xlab("Estimate")

png("data/data_analyses/models/model_plots/4l.parameters_plot_model_intervals_ABUNDANCE_LICE_brms_zinb_phylo_multiple_obs_032123_DEGREE.png",width = 3000, height = 3000, res = 300, units = "px")
estimates_plot_intervals
dev.off()

png("data/data_analyses/models/model_plots/4l.parameters_plot_model_ABUNDANCE_LICE_brms_zinb_phylo_multiple_obs_032123_DEGREE.png",width = 3000, height = 3000, res = 300, units = "px")
estimates_plot
dev.off()


# Model comparisons 
loo(zip_nf_mites_a_brms_bayes,zinb_nf_mites_a_brms_bayes,compare = TRUE)

###_###_###_###_
# Plots BRMS_w_degree
zinb_lice_a_brms_bayes_w_degree<-readRDS("data/data_analyses/models/2.model_ABUNDANCE_LICE_brms_zinb_phylo_multiple_obs_25032023_w_degree.RDS")

color_scheme_set("orange") 

# model convergence 
png("data/data_analyses/models/model_plots/4lwd.model_convergence_ABUNDANCE_LICE_brms_zinb_phylo_multiple_obs_032123_W_DEGREE.png",width = 3000, height = 3000, res = 300, units = "px")
plot(zinb_lice_a_brms_bayes_degree)
dev.off()

# model fit
png("data/data_analyses/models/model_plots/4lwd.model_fit_ABUNDANCE_LICE_brms_zinb_phylo_multiple_obs_032123_W_DEGREE.png",width = 3000, height = 3000, res = 300, units = "px")
pp_m<- brms::posterior_predict(zinb_lice_a_brms_bayes_w_degree)
ppc_rootogram(y=zinb_lice_a_brms_bayes_w_degree$data$total_lice, pp_m[1:200, ])  +   
  coord_cartesian(xlim = c(0, 100), ylim = c(0,30))
dev.off()

zinb_lice_a_brms_bayes_degree
# FOREST PLOTS OF POSTEIOR INTERVALS

estimates_plot<-mcmc_plot(zinb_lice_a_brms_bayes_w_degree,prob=0.90, prob_outer=0.95,
                          variable = c("b_Intercept", "b_scalew_degreecenterEQTRUE", "b_scaleelevation","sd_Powder.lvl__Intercept","sd_species__Intercept","sd_species_jetz__Intercept"),
                          type="areas") +
  labs(title="Posterior distribution Lice [NETWORK w_degree]", subtitle ="with medians and 95% intervals")+
  theme_minimal(20)+
  geom_vline(xintercept = 0, linetype = 2, colour = "grey40")+
  xlab("Estimate")

estimates_plot_intervals<-mcmc_plot(zinb_lice_a_brms_bayes_w_degree,prob=0.90, prob_outer=0.95,point_est = "mean",
                                    variable = c("b_Intercept", "b_scalew_degreecenterEQTRUE", "b_scaleelevation","sd_Powder.lvl__Intercept","sd_species__Intercept","sd_species_jetz__Intercept"),
                                    type="intervals") +
  labs(title="Posterior distribution Lice [NETWORK w_degree]", subtitle ="with means and 95% intervals")+
  theme_minimal(20)+
  geom_vline(xintercept = 0, linetype = 2, colour = "grey40")+
  xlab("Estimate")

png("data/data_analyses/models/model_plots/4lwd.parameters_plot_model_intervals_ABUNDANCE_LICE_brms_zinb_phylo_multiple_obs_032123_W_DEGREE.png",width = 3000, height = 3000, res = 300, units = "px")
estimates_plot_intervals
dev.off()

png("data/data_analyses/models/model_plots/4lwd.parameters_plot_model_ABUNDANCE_LICE_brms_zinb_phylo_multiple_obs_032123_W_DEGREE.png",width = 3000, height = 3000, res = 300, units = "px")
estimates_plot
dev.off()

# Model comparisons 
loo(zip_nf_mites_a_brms_bayes,zinb_nf_mites_a_brms_bayes,compare = TRUE)


# ##### 8.1 Data processing networks mites -------------------------------------------

# Mites
dff_ectos_network_individual_metrics<-read.csv("data/data_manuscript/7.dff_all_ectos_network_metrics_individuals_FILE.csv",na.strings =c("","NA"))%>% 
  select(elevation_extrapolated_date, species_jetz, Powder.lvl,foraging_cat, sociality,total_mites, total_no_feathers_mites, degree, w_degree) %>% 
  rename(elevation=elevation_extrapolated_date) %>%
  na.omit() %>% filter(species_jetz!="Premnoplex_brunnescens")  #Removing outliers for total mites


unique(dff_ectos_network_individual_metrics$species_jetz) # this is teh total species that are in flocks taht we have samples for

phylo<-read.nexus("data/phylo_data/consensus/1_consensus_birdtreeManu_ectos_prevalence.nex")  # This include speceis form manu and iquitos social and non social so we need to trim it 


# Make sure the tips and the names on the file coincide and formating of name is consitent
phylo$edge.length  
phylo$tip.label
is.binary(phylo)

# Make sure this two are the same numbers 
a<-(as.data.frame(phylo$tip.label))%>% mutate(name=phylo$tip.label) %>% select(name) %>% arrange(desc(name))
b<-(as.data.frame(dff_ectos_network_individual_metrics$species_jetz)) %>% mutate(name=dff_ectos_network_individual_metrics$species_jetz) %>% select(name) %>% arrange(desc(name)) %>% distinct(name)

tip<-as.list(setdiff(a,b))
print(tip)

# Drop some tips USE IF NEED TO DROP SOME TIPS when using the full phylogeny
phylo<-drop.tip (phylo, tip$name) 

# phylogenetic correlation structure, create a covariance matrix of species
phy_cov<-ape::vcv(phylo, corr=TRUE)

# data structure 

#dff_ectos_network_individual_metrics$elevation_cat<-as.factor(ectos_birds_dff$elevation_cat)
dff_ectos_network_individual_metrics$foraging_cat<-as.factor(dff_ectos_network_individual_metrics$foraging_cat)
dff_ectos_network_individual_metrics$species_jetz<-as.factor(dff_ectos_network_individual_metrics$species_jetz)
dff_ectos_network_individual_metrics$elevation<-as.numeric(dff_ectos_network_individual_metrics$elevation)
dff_ectos_network_individual_metrics$degree<-as.numeric(dff_ectos_network_individual_metrics$degree)
dff_ectos_network_individual_metrics$w_degree<-as.numeric(dff_ectos_network_individual_metrics$w_degree)
#ectos_birds_dff$elevation_midpoint<-as.numeric(ectos_birds_dff$elevation_midpoint)
dff_ectos_network_individual_metrics$sociality<-as.factor(dff_ectos_network_individual_metrics$sociality)
dff_ectos_network_individual_metrics$Powder.lvl<-as.factor(dff_ectos_network_individual_metrics$Powder.lvl)
#dff_ectos_network_individual_metrics$total_lice<-as.numeric(dff_ectos_network_individual_metrics$total_lice)
dff_ectos_network_individual_metrics$total_mites<-as.numeric(dff_ectos_network_individual_metrics$total_mites)
dff_ectos_network_individual_metrics$total_no_feathers_mites<-as.numeric(dff_ectos_network_individual_metrics$total_no_feathers_mites)
dff_ectos_network_individual_metrics$species<-dff_ectos_network_individual_metrics$species_jetz # create a column for the species effect different to the phylogenetic one
names(dff_ectos_network_individual_metrics)
is.ultrametric(phylo)
# ####### 8.1 Analyses models NETWORKS Mites[In progress]--------------------------------------------------------
# Mites
#Degree!!!
###_###_###


#a) model glmm
###_###_###
str(dff_ectos_network_individual_metrics)
mites_a_glmm_degree <-glmer(total_no_feathers_mites~degree+scale(elevation)+(1|Powder.lvl)+(1|species_jetz), #+elevation_midpoint+Powder.lvl
                           data = dff_ectos_network_individual_metrics, 
                           family = "poisson")

# Summary
summary(mites_a_glmm_degree )
rr2::R2(mites_a_glmm_degree)# R2 Predicted 0.38
fixef(mites_a_glmm_degree)
predict(mites_a_glmm_degree)
class(mites_a_glmm_degree) 

# Assumptions check
simulationOutput_a_mites<- DHARMa::simulateResiduals(fittedModel =mites_a_glmm_degree, plot = F,integerResponse = T, re.form = NULL ) #quantreg=T
plot(simulationOutput_a_mites)
testUniformity(simulationOutput_a_mites) #tests if the overall distribution conforms to expectations
testOutliers(simulationOutput_a_mites)#  tests if there are more simulation outliers than expected
testDispersion(simulationOutput_a_mites) # tests if the simulated dispersion is equal to the observed dispersion
testQuantiles(simulationOutput_a_mites) #fits a quantile regression or residuals against a predictor (default predicted value), and tests of this conforms to the expected quantile
testCategorical(simulationOutput_a_mites, catPred = dff_ectos_network_individual_metrics$sociality)# tests residuals against a categorical predictor
testZeroInflation(simulationOutput_a_mites) ## tests if there are more zeros in the data than expected from the simulations

# zero inflated, overdispersed, with outliers
#b) model pglmm
###_###_###

mites_a_pglmm_degree<-phyr::pglmm(total_no_feathers_mites~degree+scale(elevation)+(1|species_jetz__)+(1|Powder.lvl),
                                 data = dff_ectos_network_individual_metrics, 
                                 family = "poisson",
                                 cov_ranef = list(species_jetz= phylo), #class phylo
                                 #bayes = TRUE,
                                 add.obs.re = TRUE,
                                 REML = TRUE, 
                                 verbose = TRUE,
                                 s2.init = .25) # what is this last parameter for

histogram(dff_ectos_network_individual_metrics$total_mites) # id some outliers 
plot_data(mites_a_pglmm_degree)
summary(mites_a_pglmm_degree)
rr2::R2(mites_a_pglmm_degree)
fixef(mites_a_pglmm_degree)
predict(mites_a_pglmm_degree)

# Assumptions check
simulationOutput_a_mites_2<- DHARMa::simulateResiduals(fittedModel=mites_a_pglmm_degree, plot = F,integerResponse = T, re.form = NULL ) #quantreg=T
plot(simulationOutput_a_mites_2)
testUniformity(simulationOutput_a_mites_2) #tests if the overall distribution conforms to expectations
testOutliers(simulationOutput_a_mites_2)#  tests if there are more simulation outliers than expected
testDispersion(simulationOutput_a_mites_2) # tests if the simulated dispersion is equal to the observed dispersion
testQuantiles(simulationOutput_a_mites_2) #fits a quantile regression or residuals against a predictor (default predicted value), and tests of this conforms to the expected quantile
#testCategorical(simulationOutput_a_mites_2, catPred = dff_ectos_network_individual_metrics$sociality)# tests residuals against a categorical predictor
testZeroInflation(simulationOutput_a_mites_2) ## tests if there are more zeros in the data than expected from the simulations

# no outliers, but no homogenity of varianse KS and withig groups significant, and zero inflated

#c) model pglmm bayes: hierarchical Bayesian models fitted using integrated nested laplace approximation (INLA)
###_###_###

zip_mites_a_pglmm_bayes_degree<-phyr::pglmm(total_no_feathers_mites~scale(degree, center=TRUE)+scale(elevation)+(1|species_jetz__)+(1|Powder.lvl),
                                           data =dff_ectos_network_individual_metrics, 
                                           family ="zeroinflated.poisson", #POISSON  ="zeroinflated.poisson", #
                                           cov_ranef = list(species_jetz= phylo), #class phylo
                                           bayes = TRUE,
                                           verbose=FALSE,
                                           prior = "inla.default") # consider using    add.obs.re = T
# control = list(adapt_delta = 0.99) 
#saveRDS(zip_mites_a_pglmm_bayes_degree , "data/data_analyses/models/2.model_ABUNDANCE_nf_mites_pglmm_zip_phylo_multiple_obs_17032023_degree.RDS")
zip_mites_a_pglmm_bayes<-readRDS ("data/data_analyses/models/2.model_ABUNDANCE_nf_mites_pglmm_zip_phylo_multiple_obs_17032023_degree.RDS")
#1) Summary of the model 
summary(zip_mites_a_pglmm_bayes_degree)   
rr2::R2(zip_mites_a_pglmm_bayes_degree)
class(zip_mites_a_pglmm_bayes_degree)

launch_shinystan(zip_mites_a_pglmm_bayes_degree)

# Plots
estimates_plot<-plot_bayes(zip_mites_a_pglmm_bayes_degree ) # for some reason does not allow me to plot zero inflated poisson 
#png("data/data_analyses/models/model_plots/2.parameters_plot_model_mites_ABUNDANCE_pglmm_zip_phylo_multiple_obs_032123.png",width = 3000, height = 3000, res = 300, units = "px")
#estimates_plot
#dev.off()

# Assumptions check DHARMa does not work with pglm bayes=TRUE so I can not evaluate model fit.
simulationOutput_a_mites_3<- DHARMa::simulateResiduals(fittedModel=zip_mites_a_pglmm_bayes_degree, plot = F,integerResponse = T, re.form = NULL ) #quantreg=T
plot(simulationOutput_a_mites_3)
testUniformity(simulationOutput_a_mites_3) #tests if the overall distribution conforms to expectations
testOutliers(simulationOutput_a_mites_3)#  tests if there are more simulation outliers than expected
testDispersion(simulationOutput_a_mites_3) # tests if the simulated dispersion is equal to the observed dispersion
testQuantiles(simulationOutput_a_mites_3) #fits a quantile regression or residuals against a predictor (default predicted value), and tests of this conforms to the expected quantile
testCategorical(simulationOutput_a_mites_3, catPred = dff_ectos_network_individual_metrics$sociality)# tests residuals against a categorical predictor
testZeroInflation(simulationOutput_a_mites_3) ## tests if there are more zeros in the data than expected from the simulations


#d) model brms bayes : zero_inflated_negbinomial, ind scaled elevation
###_###_###

improved_prior <- prior(normal(0, 10), class = "b") + 
  prior(student_t(3, 0, 10), class = "Intercept") + 
  prior(normal(0, 10), class = "sd") 

# degree zinb

zinb_nf_mites_a_brms_bayes_degree<-brm(total_no_feathers_mites~scale(degree, center=TRUE) +scale(elevation)+
                                     (1|gr(species_jetz, cov = phy_cov))+  
                                     (1|Powder.lvl) + 
                                     (1|species),
                                   data=dff_ectos_network_individual_metrics,
                                   family=zero_inflated_negbinomial(),  #zero_inflated_negbinomial()
                                   data2 = list(phy_cov=phy_cov),
                                   iter=8000, warmup=4000,thin=4,
                                   control=list(adapt_delta=0.99, max_treedepth=12)) 
#                            prior = improved_prior,


#saveRDS(zinb_nf_mites_a_brms_bayes_degree, "data/data_analyses/models/2.model_ABUNDANCE_nf_mites_brms_zinb_phylo_multiple_obs_25032023_degree.RDS")
zinb_nf_mites_a_brms_bayes_degree<-readRDS("data/data_analyses/models/2.model_ABUNDANCE_nf_mites_brms_zinb_phylo_multiple_obs_25032023_degree.RDS")

hypothesis(zinb_mites_a_brms_bayes_degree,"scale(degree)>0", alpha=0.05 ) # increase of degree increases mites abundance
hypothesis(zinb_mites_a_brms_bayes_degree,"total_mites=degree+species", alpha=0.05 ) # increase of degree increases mites abundance

#  exploring zero inflated poisson 
# degree zip

zip_mites_a_brms_bayes_degree<-brm(total_no_feathers_mites~scale(degree, center=TRUE) +scale(elevation)+
                                    (1|gr(species_jetz, cov = phy_cov))+  
                                    (1|species),
                                  data=dff_ectos_network_individual_metrics,
                                  family=zero_inflated_poisson(),  #zero_inflated_negbinomial()
                                  data2 = list(phy_cov=phy_cov),
                                  iter=8000, warmup=4000,
                                  thin=4,
                                  control=list(adapt_delta=0.99, max_treedepth=12)) 

zip_mites_a_brms_bayes_degree
#saveRDS(zip_mites_a_brms_bayes_degree, "data/data_analyses/models/2.model_ABUNDANCE_nf_mites_brms_zip_phylo_multiple_obs_21032023_degree.RDS")

# some outliers and smaller R2, overall performed porrly compared to zeroinflated negative binomial)_

# mites
#w_Degree!!!

zinb_nf_mites_a_brms_bayes_w_degree<-brm(total_no_feathers_mites~scale(w_degree, center=TRUE) +scale(elevation)+
                                       (1|gr(species_jetz, cov = phy_cov))+  
                                       (1|Powder.lvl) + 
                                       (1|species),
                                     data=dff_ectos_network_individual_metrics,
                                     family=zero_inflated_negbinomial(),  #zero_inflated_negbinomial()
                                     data2 = list(phy_cov=phy_cov),
                                     iter=8000, warmup=4000,thin=4,
                                     #prior=improved_prior,
                                     control=list(adapt_delta=0.99, max_treedepth=12)) 


#saveRDS(zinb_nf_mites_a_brms_bayes_w_degree, "data/data_analyses/models/2.model_ABUNDANCE_nf_mites_brms_zinb_phylo_multiple_obs_25032023_w_degree.RDS")
zinb_nf_mites_a_brms_bayes_w_degree<-readRDS("data/data_analyses/models/2.model_ABUNDANCE_nf_mites_brms_zinb_phylo_multiple_obs_25032023_w_degree.RDS")

hypothesis(zinb_mites_a_brms_bayes_degree,"degree>0", alpha=0.05 ) # increase of degree increases mites abundance
hypothesis(zinb_mites_a_brms_bayes_degree,"total_mites=degree+species", alpha=0.05 ) # increase of degree increases mites abundance

#  exploring zero inflated poisson 

zip_nf_mites_a_brms_bayes_w_degree=zip_mites_a_brms_bayes_w_degree
zip_mites_a_brms_bayes_w_degree<-brm (total_no_feathers_mites~scale(w_degree,center=TRUE)+scale(elevation)+
                                       (1|gr(species_jetz, cov = phy_cov))+  
                                       (1|species),
                                     data=dff_ectos_network_individual_metrics,
                                     family=zero_inflated_poisson(),  #zero_inflated_negbinomial()
                                     data2 = list(phy_cov=phy_cov),
                                     iter=8000, warmup=4000,
                                     thin=2,
                                     control=list(adapt_delta=0.999, max_treedepth=12)) 


#saveRDS(zip_nf_mites_a_brms_bayes_w_degree, "data/data_analyses/models/2.model_ABUNDANCE_nf_mites_brms_zip_phylo_multiple_obs_21032023_w_degree.RDS")


# Summarize the models
summary(zinb_mites_a_brms_bayes_degree)
fixef(zinb_mites_a_brms_bayes_degree) # to get more detailed values for estimates
coef(zinb_mites_a_brms_bayes_degree) # if you have group-level effects (hierarchical data)

# interpret the model 
#To test whether all regression coefficients are different from zero, we can look at the Credible Intervals that are listed in the summary output or we can visually represent them in density plots.
#If we do so, we clearly see that zero is not included in any of the density plots, meaning that we can be reasonably certain the regression coefficients are different from zero.
#INTERPRETATION:In the model, the parameter for Sociality means the expected difference between non_social(0) and social (1) with all other covariates held constant. we clearly see that zero is included in the density plot for sociality so there is not effect of sociality??
bayes_R2(zinb_nf_mites_a_brms_bayes_degree) # zing 0.31
plot(zinb_mites_a_brms_bayes_degree)
mcmc_plot(zinb_nf_mites_a_brms_bayes_w_degree) # Dots represent means of posterior distribution along with 95% CrIs, as estimated by the bmod5 model
launch_shinystan()
pp_check(zinb_mites_a_brms_bayes_degree, ndraws = 100)+ xlim(0, 5)  #  test for the model fit to the data .need to modify the scale of this plot posterior predictive checks, 100 random draws or distributions created by the model 
pp_check(zinb_mites_a_brms_bayes_degree, type="bars", ndraws = 100)+ xlim(0, 20) 

pp_m<- brms::posterior_predict(zinb_mites_a_brms_bayes_degree)
log1 <- scale_x_continuous(trans="log1p")
ppc_dens_overlay(y=zinb_mites_a_brms_bayes_degree$data$total, pp_m[1:200, ])  + 
  coord_cartesian(xlim = c(0, 5))

pp_m<- brms::posterior_predict(zinb_mites_a_brms_bayes_degree)
ppc_rootogram(y=zinb_mites_a_brms_bayes_degree$data$total_mites, pp_m[1:200, ])  +   
  coord_cartesian(xlim = c(0, 5), ylim = c(0,30))

ppc_stat(y=zinb_mites_a_brms_bayes_degree$data$total_mites, pp_m, stat ="prop_zero")


#Assumptions check model pglmm_ bayes
#remotes::install_github("Pakillo/DHARMa.helpers")

simulate_residuals <- dh_check_brms(zinb_mites_a_brms_bayes_degree, integer = TRUE)
plot( simulate_residuals, form = dff_ectos_network_individual_metrics$sociality)
DHARMa::testDispersion(simulate_residuals)
DHARMa::testZeroInflation(simulate_residuals ) ## tests if there are more zeros in the data than expected from the simulations
testUniformity(simulate_residuals) #tests if the overall distribution conforms to expectations

# plots BRMS degree mites
zinb_nf_mites_a_brms_bayes_degree<-readRDS("data/data_analyses/models/2.model_ABUNDANCE_nf_mites_brms_zinb_phylo_multiple_obs_25032023_DEGREE.RDS")
color_scheme_set("red") 

# model convergence 
png("data/data_analyses/models/model_plots/4md.model_convergence_ABUNDANCE_nf_MITES_brms_zinb_phylo_multiple_obs_032123_DEGREE.png",width = 3000, height = 3000, res = 300, units = "px")
plot(zinb_nf_mites_a_brms_bayes_degree)
dev.off()

# model fit
png("data/data_analyses/models/model_plots/4md.model_fit_ABUNDANCE_nf_MITES_brms_zinb_phylo_multiple_obs_032123_DEGREE.png",width = 3000, height = 3000, res = 300, units = "px")
pp_m<- brms::posterior_predict(zinb_nf_mites_a_brms_bayes_degree)
ppc_rootogram(y=zinb_nf_mites_a_brms_bayes$data$total_no_feathers_mites, pp_m[1:200, ])  +   
  coord_cartesian(xlim = c(0, 100), ylim = c(0,30))
dev.off()


# FOREST PLOTS OF POSTEIOR INTERVALS

estimates_plot<-mcmc_plot(zinb_nf_mites_a_brms_bayes_degree,prob=0.90, prob_outer=0.95,
                          variable = c("b_Intercept", "b_scaledegreecenterEQTRUE", "b_scaleelevation","sd_Powder.lvl__Intercept","sd_species__Intercept","sd_species_jetz__Intercept"),
                          type="areas") +
  labs(title="Posterior distribution Mites (n_f)[Networks_degree]", subtitle ="with medians and 95% intervals")+
  theme_minimal(20)+
  geom_vline(xintercept = 0, linetype = 2, colour = "grey40")+
  xlab("Estimate")

estimates_plot_intervals<-mcmc_plot(zinb_nf_mites_a_brms_bayes_degree,prob=0.90, prob_outer=0.95,point_est = "mean",
                                    variable = c("b_Intercept", "b_scaledegreecenterEQTRUE", "b_scaleelevation","sd_Powder.lvl__Intercept","sd_species__Intercept","sd_species_jetz__Intercept"),
                                    type="intervals") +
  labs(title="Posterior distribution Mites(n_f)[Networks_degree]", subtitle ="with means and 95% intervals")+
  theme_minimal(20)+
  geom_vline(xintercept = 0, linetype = 2, colour = "grey40")+
  xlab("Estimate")

png("data/data_analyses/models/model_plots/4md.parameters_plot_intervals_model_intervals_ABUNDANCE_nf_MITES_brms_zinb_phylo_multiple_obs_032123_DEGREE.png",width = 3000, height = 3000, res = 300, units = "px")
estimates_plot_intervals
dev.off()

png("data/data_analyses/models/model_plots/4md.parameters_plot_model_ABUNDANCE_MITES_brms_zinb_phylo_multiple_obs_032123_DEGREE.png",width = 3000, height = 3000, res = 300, units = "px")
estimates_plot
dev.off()

# Model comparisons 
loo(zinb_nf_mites_a_brms_bayes_w_degree,zip_mites_a_brms_bayes_w_degree,compare = TRUE)

# Plots BRMS_w_degree_mites
#W_degree
zinb_nf_mites_a_brms_bayes_w_degree<-readRDS("data/data_analyses/models/2.model_ABUNDANCE_nf_mites_brms_zinb_phylo_multiple_obs_25032023_w_degree.RDS")
color_scheme_set("purple") 

# model convergence 
png("data/data_analyses/models/model_plots/4mwd.model_convergence_ABUNDANCE_MITES_brms_zinb_phylo_multiple_obs_032123_W_DEGREE.png",width = 3000, height = 3000, res = 300, units = "px")
plot(zinb_nf_mites_a_brms_bayes_w_degree)
dev.off()

# model fit
png("data/data_analyses/models/model_plots/4mwd.model_fit_ABUNDANCE_MITES_brms_zinb_phylo_multiple_obs_032123_W_DEGREE.png",width = 3000, height = 3000, res = 300, units = "px")
pp_m<- brms::posterior_predict(zinb_nf_mites_a_brms_bayes_w_degree)
ppc_rootogram(y=zinb_nf_mites_a_brms_bayes_w_degree$data$total_no_feathers_mites, pp_m[1:200, ])  +   
  coord_cartesian(xlim = c(0, 100), ylim = c(0,30))
dev.off()

# FOREST PLOTS OF POSTEIOR INTERVALS

estimates_plot<-mcmc_plot(zinb_nf_mites_a_brms_bayes_w_degree,prob=0.90, prob_outer=0.95,
                          variable = c("b_Intercept", "b_scalewdegreecenterEQTRUE", "b_scaleelevation","sd_Powder.lvl__Intercept","sd_species__Intercept","sd_species_jetz__Intercept"),
                          type="areas") +
  labs(title="Posterior distribution Mites [NETWORK_w_degree]", subtitle ="with medians and 95% intervals")+
  theme_minimal(20)+
  geom_vline(xintercept = 0, linetype = 2, colour = "grey40")+
  xlab("Estimate")

estimates_plot_intervals<-mcmc_plot(zinb_mites_a_brms_bayes_w_degree,prob=0.90, prob_outer=0.95,point_est = "mean",
                                    variable = c("b_Intercept", "b_scalewdegreecenterEQTRUE", "b_scaleelevation","sd_Powder.lvl__Intercept","sd_species__Intercept","sd_species_jetz__Intercept"),
                                    type="intervals") +
  labs(title="Posterior distribution Mites [NETWORK_w_degree]", subtitle ="with means and 95% intervals")+
  theme_minimal(20)+
  geom_vline(xintercept = 0, linetype = 2, colour = "grey40")+
  xlab("Estimate")

png("data/data_analyses/models/model_plots/4mwd.parameters_plot_model_intervals_ABUNDANCE_MITES_brms_zinb_phylo_multiple_obs_032123_W_DEGREE.png",width = 3000, height = 3000, res = 300, units = "px")
estimates_plot_intervals
dev.off()

png("data/data_analyses/models/model_plots/4mwd.parameters_plot_model_ABUNDANCE_MITES_brms_zinb_phylo_multiple_obs_032123_W_DEGREE.png",width = 3000, height = 3000, res = 300, units = "px")
estimates_plot
dev.off()

# #MODIFYING SOME PLOTS FOR BAYESIAN MODELS [FUNCTIONS] ----------------------------------


#### playing with teh functions to create plots for bayesian

# i MODIFIED THE FUNCTION TO BE ABLE TO PLOT IT IN NEGATIVE BINOMIAL


plot_bayes.communityPGLMM <- function(x, n_samp = 1000, sort = TRUE, ...) {
  
  if(!requireNamespace("ggplot2", quietly = TRUE)) {
    stop('plot_bayes requires the ggplot2 package but it is unavailable. Use install.packages("ggplot2") to install it.')
  }
  
  if(!x$bayes) {
    stop("plot_bayes only works on communityPGLMM objects fit with bayes = TRUE")
  }
  
  if(!requireNamespace("ggridges", quietly = TRUE)) {
    stop('plot_bayes requires the ggridges package but it is unavailable. Use install.packages("ggridges") to install it.')
  }
  
  r.effect_jen<-c("1|species_jetz", "1|species_jetz__","1|Powder.lvl","1|obs")
  
  re.names <-r.effect_jen
  if (x$family == "zeroinflated.poisson") re.names <- c("residual", re.names)
  random_samps <- lapply(x$inla.model$marginals.hyperpar, 
                         function(x) INLA::inla.rmarginal(1000, INLA::inla.tmarginal(function(x) sqrt(1 / x), x))) %>%
    setNames(re.names) %>%
    dplyr::as_tibble() %>%
    tidyr::pivot_longer(cols = dplyr::everything(),
                        names_to = "var",
                        values_to = "val") %>%
    dplyr::mutate(effect_type = "Random Effects")
  
  fixed_samps <- lapply(x$inla.model$marginals.fixed, function(x) INLA::inla.rmarginal(1000, x)) %>%
    dplyr::as_tibble() %>%
    tidyr::pivot_longer(cols = dplyr::everything(),
                        names_to = "var",
                        values_to = "val") %>%
    dplyr::mutate(effect_type = "Fixed Effects")
  
  samps <- dplyr::bind_rows(random_samps, fixed_samps) %>%
    dplyr::mutate(effect_type = factor(effect_type, 
                                       levels = c("Random Effects", "Fixed Effects")))
  
  ci <- samps %>%
    dplyr::group_by(var, effect_type) %>%
    dplyr::summarise(lower = quantile(val, 0.025),
                     upper = quantile(val, 0.975),
                     mean = mean(val),
                     .groups = "drop_last")
  
  if(sort){
    ci <- dplyr::arrange(ci, mean) %>% dplyr::ungroup() %>% 
      dplyr::mutate(var = factor(as.character(var), levels = as.character(var)))
  }
  
  sig_vars <- ci %>%
    dplyr::mutate(sig = ifelse(effect_type == "Random Effects",
                               "CI no overlap with zero",
                               ifelse(sign(lower) == sign(upper),
                                      "CI no overlap with zero",
                                      "CI overlaps zero"))) %>%
    dplyr::select(var, sig)
  
  if(sort){
    samps <- dplyr::mutate(samps, var = factor(var, levels = levels(sig_vars$var)))
  }
  
  samps <- samps %>%
    dplyr::left_join(sig_vars, by = "var") %>%
    dplyr::group_by(var) %>%
    dplyr::filter(abs(val - mean(val)) < (10 * sd(val))) %>% 
    dplyr::ungroup()
  
  pal <- c("#fc8d62", "#8da0cb")
  
  #samps<-samps %>% filter(var!="1|obs")
  p <- ggplot2::ggplot(samps, ggplot2::aes(val, var, height = ..density..)) +
    ggridges::geom_density_ridges(ggplot2::aes(alpha = sig, fill = sig), 
                                  stat = "density", adjust = 2, color = "gray70") +
    ggplot2::geom_point(ggplot2::aes(x = mean, y = var), data = ci, inherit.aes = FALSE) +
    ggplot2::geom_errorbarh(ggplot2::aes(xmin = lower, xmax = upper, y = var), data = ci,
                            inherit.aes = FALSE, height = 0.1) +
    ggplot2::geom_vline(xintercept = 0, linetype = 2, colour = "grey40") +
    ggplot2::scale_alpha_manual(values = c(0.8, 0.2)) +
    ggplot2::scale_fill_manual(values = rev(pal)) +
    ggplot2::facet_wrap(~ effect_type, nrow = 2, scales = "free") +
    ggplot2::ylab("") +
    ggplot2::xlab("Estimate") +
    ggplot2::theme_minimal() +
    ggplot2::theme(legend.position = "none",
                   axis.text = ggplot2::element_text(size = 14),
                   strip.text = ggplot2::element_text(size = 16))
  
  p
}

##### Note on interpreting the plots of the posterior distributions with credible intervals
#Remember that the posterior distribution represents our uncertainty (or certainty) in "q"
#after combining the information in the data (the likelihood) with what we knew before collecting data (the prior).
#Bayesian inference we quantify statements like this – that a particular event is “highly likely” – by computing the “posterior probability” of the event,
#which is the probability of the event under the posterior distribution.

#This plots do not tell you the significance of the factors, instead they tell you the effect size of the factors.
# The most common summaries of a posterior distribution are interval estimates and point estimates.
# Point estimates are typically obtained by computing the mean or median (or mode) of the posterior distribution. These are called the “posterior mean” or the “posterior median” (or “posterior mode”).
# Point estimates Essentially this boils down to summarizing the posterior distribution by a single number.
# Interval estimates can be obtained by computing quantiles of the posterior distribution. Bayesian Confidence intervals are often called “Credible Intervals”.
# We can extend this idea to assess the certainty (or confidence) that "q" lies in any interval. For example, from the plot it looks like q
# will very likely lie in the interval [0.2,0.4] because most of the posterior distribution mass lies between these two numbers.
#it is more common to compute Bayesian Confidence Intervals the other way around: specify the level of confidence we want to achieve and find an interval that achieves that level of confidence. 
#This can be done by computing the quantiles of the posterior distribution. For example, the 0.05 and 0.95 quantiles of the posterior would define a 90% Bayesian Confidence Interval.
# Positive estimates are positive effects and  negative values  negative effects, when they overlap with zero you can conclude the effect size does not differ from zero.

# In our models the predictor sociality has two levels 0 sociality and 1 sociality, and so
  #!. The intercept estimate  CORRRESPOND TO  the first  level of sociality ( sociality zero)
  # sociality estimate is the difference between the sociality zero and the sociality one, so if it overlaps with zero means teh difference is close to zero between being social and non social  

# notes with BRSM fitting 

#The brm() version of the GAM is fitted using the code below. Note that I have changed a few things from their default values as

#the model required more than the default number of MCMC samples — iter = 4000,
#the samples needed thinning to deal with some strong autocorrelation in the Markov chains — thin = 10,
#the adapt.delta parameter, a tuning parameter in the NUTS sampler for Hamiltonian Monte Carlo, potentially needed raising — there was a warning about a potential divergent transition but I should have looked to see if it was one or not; instead I just increased the tuning parameter to 0.99,
#four chains fitted by default but I wanted these to be fitted using 4 CPU cores,
#seed sets the internal random number generator seed, which allows reproducibility of models, and
#for this post I didn’t want to print out the progress of the sampler — refresh = 0 — typically you won’t want to do this so you can see how sampling is progressing.
